### 2017-01-11 #161 @whedon
[REVIEW]: adaptest: Data-Adaptive Statistics for High-Dimensional Testing in R
**Submitting author:** @wilsoncai1992 (<a href="http://orcid.org/0000-0003-2680-3066">Weixin Cai</a>)
**Repository:** <a href="https://github.com/wilsoncai1992/adaptest" target ="_blank">https://github.com/wilsoncai1992/adaptest</a>
**Version:** 1.1.0

- [x] **Authorship:** Has the submitting author (wilsoncai1992) made major contributions to the software?
