### 2017-01-11 850c0c0
fix travis, reformat readme
