### 2017-01-11 5f506ea
fix S3 methods, rebuild pkgdown site
