+++ pkg/README.md
@@ -1,11 +1,12 @@
 # R/`data.adapt.multi.test`
 
-[![Travis-CI Build
-Status](https://travis-ci.org/wilsoncai1992/data.adapt.multi.test.svg?branch=master)](https://travis-ci.org/wilsoncai1992/data.adapt.multi.test?branch=master)
-[![Coverage
-Status](https://coveralls.io/repos/github/wilsoncai1992/data.adapt.multi.test/badge.svg?branch=master)](https://coveralls.io/github/wilsoncai1992/data.adapt.multi.test?branch=master)
-[![GPL-2
-License](http://img.shields.io/:license-gpl2-blue.svg)](http://www.gnu.org/licenses/gpl-2.0.html)
+[![Travis-CI Build Status](https://travis-ci.org/wilsoncai1992/data.adapt.multi.test.svg?branch=master)](https://travis-ci.org/wilsoncai1992/data.adapt.multi.test?branch=master)
+[![Coverage Status](https://coveralls.io/repos/github/wilsoncai1992/data.adapt.multi.test/badge.svg?branch=master)](https://coveralls.io/github/wilsoncai1992/data.adapt.multi.test?branch=master)
+[![CRAN](http://www.r-pkg.org/badges/version/data.adapt.multi.test)](http://www.r-pkg.org/pkg/data.adapt.multi.test)
+[![Project Status: Active - The project has reached a stable, usable state and is being actively developed.](http://www.repostatus.org/badges/latest/active.svg)](http://www.repostatus.org/#active)
+[![GPL-2 License](http://img.shields.io/:license-gpl2-blue.svg)](http://www.gnu.org/licenses/gpl-2.0.html)
+[![JOSS status](http://joss.theoj.org/papers/7618d7d14ac77f6f502df3f9eac5917d/status.svg)](http://joss.theoj.org/papers/7618d7d14ac77f6f502df3f9eac5917d)
+
 
 > Data-adaptive statistics for multiple testing in high-dimensional problems.
 
@@ -31,10 +32,21 @@ data-adaptive statistical target parameters and inference.
 
 ## Citation
 
-To cite the `data.adapt.multi.test` R package in publications, please use:
-> Cai W., Hejazi N., Hubbard A. (2016). *data.adapt.multi.test*: Data-Adaptive
-> Statistics for High-Dimensional Multiple Testing. *The Journal of Open Source
-> Software*.
+After using the `data.adapt.multi.test` R package, please cite it:
+
+      @article{cai2017,
+        doi = {},
+        url = {},
+        year  = {2017},
+        month = {},
+        publisher = {The Open Journal},
+        volume = {},
+        number = {},
+        author = {Cai, Weixin and Hejazi, Nima and Hubbard, Alan},
+        title = {data.adapt.multi.test: Data-Adaptive Statistics for
+          High-Dimensional Testing},
+        journal = {The Journal of Open Source Software}
+      }
 
 ---
 
@@ -48,4 +60,5 @@ To cite the `data.adapt.multi.test` R package in publications, please use:
 
 ## License
 
-This software is distributed under the GPL-2 license.
+This software contents of this repository are distributed under the GPL-2
+license. See file `LICENSE` for details.
