+++ pkg/NAMESPACE
@@ -1,10 +1,10 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(plot,data_adapt)
+S3method(print,data_adapt)
 export(data_adapt)
 export(data_adapt_multi_test)
 export(data_adapt_rank)
-export(plot.data_adapt)
-export(print.data_adapt)
 export(shinyprint.data_adapt)
 export(table.data_adapt)
 importFrom(R2HTML,HTML)
