### 2017-01-12 949b395
add working paper, fix JOSS citation
