### 2017-03-14 d5320f5
fix plot method
