+++ pkg/man/data_adapt_multi_test.Rd
@@ -39,4 +39,3 @@ associated with multiple testing.
 Wilson Cai \email{wcai@berkeley.edu}, in collaboration with Alan E.
         Hubbard, with contributions from Nima S. Hejazi.
 }
-
