+++ pkg/DESCRIPTION
@@ -26,5 +26,4 @@ URL:
     https://github.com/wilsoncai1992/data.adapt.multi.test
 LazyData:
     true
-RoxygenNote:
-    5.0.1
+RoxygenNote: 6.0.0
