### 2017-03-20 wilsoncai1992/data.adapt.multi.test#11 @gavinsimpson
JOSS review: References should include DOIs where cited works have them
The Hubbard et al paper in the references should have the a DOI (10.1515/ijb-2015-0013) listed alongside the reference metadata to facilitate locating this work. openjournals/joss-reviews#161
