### 2017-03-20 wilsoncai1992/adaptest#12 @gavinsimpson
Add examples to R documentation (JOSS review)
User-facing functions in an R package should have an Examples section indicating how the function is used and some of the functionality. In addition, packages authors might include a vignette which acts as documentation, often in tutorial form, that illustrates in a practical sense how the package works.

I can't sign off on this element (openjournals/joss-reviews#161):

> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

currently as none of the Rd pages have examples and/or there is no vignette or other associated documentation that fills this criterion.
