### 2017-03-21 wilsoncai1992/data.adapt.multi.test#12 @wilsoncai1992
Dear @gavinsimpson,

Thank you so much for pointing out! We have some tutorials at hand and will start putting up a vignette immediately.
