### 2017-04-08 wilsoncai1992/data.adapt.multi.test#12 @gavinsimpson
It requires more than a vignette; Writing R Extensions, as well as conventions in the R world require that user-facing functions should all have an `\examples{}` section in the Rd file with example code that illustrates how to use said function.
