### 2017-04-08 wilsoncai1992/data.adapt.multi.test#13 @gavinsimpson
JOSS Paper - more clearly state what deficiencies with multiple testing
The paper currently alludes to problems with existing multiple testing methods, particularly in high dimensional settings, but fails to explain, even briefly, what these deficiencies are.

Please state more clearly what the problems are that **data.adapt.multi.test** solves, not just that there are problems.

This is needed to address the **Statement of Need** sections of the JOSS review.

openjournals/joss-reviews#161
