### 2017-04-08 wilsoncai1992/data.adapt.multi.test#14 @gavinsimpson
R CMD check NOTE re an extraneous dependency in Imports
When checking the pkg I see the following

> * checking dependencies in R code ... NOTE
> Namespace in Imports field not imported from: ‘SuperLearner’
>  All declared Imports should be used.

If you are not importing from **SuperLearner**, please remove it from the `Imports:` field in `DESCRIPTION`.

openjournals/joss-reviews/issues/161
