### 2017-04-08 #161 @gavinsimpson
If you intend to submit this to CRAN, change the `Title` field to read:

    Data-Adaptive Test Statistics for Multiple Testing in High Dimensions

i.e. in what CRAN calls Title Case.
