### 2017-04-08 wilsoncai1992/DA.Test#15 @gavinsimpson
JOSS Review: Consider adding information about contributing and guidelines for doing so
In order to clearly indicate how users can contribute to the development of the package and the guidelines for doing so, please consider adding a `CONTRIBUTING` (or `CONTRIBUTING.md`) file to the top level of the repo.

openjournals/joss-reviews/issues/161
