### 2017-04-08 wilsoncai1992/adaptest#16 @gavinsimpson
JOSS Review: unit tests should cover all of the package functions
The only unit tests in the package currently current is a single test of the `data_adapt()` function, which tests if the function fails if called incorrectly. As such there are no tests to confirm the functions in the package work as expected.

Please add unit tests that test all functions in the package, especially that the functions work as intended, rather than only failing with incorrect inputs.

openjournals/joss-reviews/issues/161
