### 2017-04-08 #161 @gavinsimpson
At this point, the main issue preventing finalisation of this review is the general lack of examples and documentation that I can run to explore the functionality of the package.

I have filed a number of issues for the outstanding elements of the review check list. Once examples have been added to user-facing functions and the Vignette the authors indicate they are writing, I can tackle the final outstanding review topic of Functionality.

Please ping me when the filed issues have been addressed or responded to.
