### 2017-04-08 wilsoncai1992/DA.Test#18 @gavinsimpson
JOSS Review: Provide .Rmd source for the JOSS paper
The Rmd or other source for the JOSS paper would be useful to see how the pkg was used in the simulated data example described in the JOSS paper.

openjournals/joss-reviews#161
