### 2017-04-21 wilsoncai1992/data.adapt.multi.test#14 @nhejazi
We've not been able to replicate the problem stated in this issue locally. It's possible that it was fixed as a side-effect of recent changes. We'll continue trying and post updates as necessary.
