### 2017-04-21 wilsoncai1992/data.adapt.multi.test#11 @nhejazi
fixed in 089ff36
