### 2017-04-21 wilsoncai1992/data.adapt.multi.test#13 @wilsoncai1992
The issue is completed by commit 840ace169b351976d4e5d3435ef98a345c4e2116
