### 2017-04-25 wilsoncai1992/data.adapt.multi.test#14 @gavinsimpson
Interestingly, I don't see this if I do `R CMD check --as-cran`.

However, the problem is simple; you have `SuperLearner` in `Imports` (in `DESCRIPTION` in the master branch: https://github.com/wilsoncai1992/data.adapt.multi.test/blob/master/pkg/DESCRIPTION#L17) and yet you don't use this package anywhere (I grepped the R source to check) and don't use it in any `@importFrom` statements.

If you're not using this package in your package then just remove this line from `DESCRIPTION`. If you've already fixed this in the development branch (which I haven't checked) then you're good and this can be closed.
