### 2017-04-25 73330f9
remove SuperLearner from Imports. Fixes #14
