### 2017-04-25 wilsoncai1992/data.adapt.multi.test#14 @nhejazi
I've seen this problem come up in other packages I've worked on before, and yet I always forget exactly what the fix ends up being.

There is a problem with removing `SuperLearner` from `Imports` -- specifically, I think it gets re-added by `devtools::document` because `tmle::tmle` calls `SuperLearner::SuperLearner` internally.

It is rather odd that this seems to slip past `R CMD check --as-cran`.

I'll look into this further and report back if/when I am able to find a suitable fix.
