### 2017-04-25 wilsoncai1992/data.adapt.multi.test#14 @gavinsimpson
What **tmle** does should be of no consequence for your package. You might want to ask if this is intended behaviour of `devtools::document()` as this is not what I would expect to happen nor what I understand the NAMESPACE rules to be.

This may be an issue in roxygen of course, as `document()` is just a wrapper to it.
