### 2017-05-24 wilsoncai1992/DA.Test#15 @nhejazi
This was completed by 089ff36, which, among other changes, added `CONTRIBUTING.md`
