### 2017-05-24 wilsoncai1992/DA.Test#18 @nhejazi
The JOSS paper had merely been written in markdown, with figures generated and added manually.  An Rmd file, which will contain R code for producing the figures, was added to the `paper` directory in a903762.
