+++ sim/data_adaptive_simulate.R
@@ -1,11 +1,15 @@
 set.seed(4321)
 n.sim <- 1e2 # sample size
+# n.sim <- 1e3 # sample size
 # n.sim <- 1 # sample size
-p.all <- 1e3 # No. of dimensions of Y (signal + non-signal)
+# p.all <- 1e3 # No. of dimensions of Y (signal + non-signal)
+p.all <- 1e4 # No. of dimensions of Y (signal + non-signal)
+# p.all <- 1e5 # No. of dimensions of Y (signal + non-signal)
 # p.all <- 1e6 # No. of dimensions of Y (signal + non-signal)
 p.true <- 10 # No. of dimensions for signal
-signal.true <- 1 # size of true signal
-# signal.true <- .5 # size of true signal
+# signal.true <- 1 # size of true signal
+signal.true <- .6 # size of true signal
+# signal.true <- .8 # size of true signal
 
 # n.top.want <- p.true
 n.top.want <- 15
@@ -53,13 +57,28 @@ gc()
 # ------------------------------------
 
 # ==============================================================================
-# fit model
+# fit DA.test
 # ==============================================================================
 library(data.adapt.multi.test)
 out_result <- data_adapt_multi_test(Y = Y, A = A.sample.vec, n.top = p.true + 5,
-                                    n.fold = 10, parallel = FALSE) # BH on all Y
-
-print.data_adapt(out_result)
-plot.data_adapt(out_result)
+                                    # n.fold = 3, parallel = FALSE) # BH on all Y
+																		n.fold = 4, parallel = FALSE) # BH on all Y
+print(out_result)
+plot(out_result)
+# print.data_adapt(out_result)
+# plot.data_adapt(out_result)
 
 save(out_result, file = 'out_result_1e3_sig_1.Rdata')
+# ==============================================================================
+# fit naive BH
+# ==============================================================================
+lm_out <- lm(Y ~ A.sample.vec)
+B1_hat_lm <- lm_out$coefficients[2,]
+lm_summary <- summary(lm_out)
+pval_lm <- sapply(lm_summary, function(x) x$coefficients[2,4])
+
+qval_lm <- p.adjust(pval_lm, method = 'BH')
+which(qval_lm < .05)
+head(sort(qval_lm), 20)
+plot(head(sort(qval_lm), 100))
+plot(sort(qval_lm))
