### 2017-10-05 a768901
hard override of master branch
