+++ pkg/R/data_adapt_multi_test.R
@@ -77,7 +77,9 @@ data_adapt_multi_test <- function(Y,
   folds.table <- table(n.index.param.gen)
 
   rank.all.fold <- matrix(0, nrow = n.fold, ncol = p.all)
-
+  adapt_param_composition <- matrix(0, nrow = n.fold, ncol = n.top)
+  psi.est_composition <- list()
+  EIC.est_composition <- list()
   # ============================================================================
   compute.a.fold <- function(data_adapt, it0) {
     print(paste('Fold:', it0))
@@ -86,38 +88,56 @@ data_adapt_multi_test <- function(Y,
     # create parameter generating data
     Y.param <- data_adapt$Y[n.index.param.gen != chunk.as.est, ]
     A.param <- data_adapt$A[n.index.param.gen != chunk.as.est]
-    W.param <- data_adapt$W[n.index.param.gen != chunk.as.est, drop = FALSE]
+    W.param <- data_adapt$W[n.index.param.gen != chunk.as.est, ,drop = FALSE]
 
     # create estimation data
     Y.est <- data_adapt$Y[n.index.param.gen == chunk.as.est, ]
     A.est <- data_adapt$A[n.index.param.gen == chunk.as.est]
-    W.est <- data_adapt$W[n.index.param.gen == chunk.as.est, drop = FALSE]
+    W.est <- data_adapt$W[n.index.param.gen == chunk.as.est, ,drop = FALSE]
 
-    # ==========================================================================
+    #
     # data-adaptive target parameter
-    # ==========================================================================
+    # ---------------------------------------------------------------------------------------
     data.adaptive.index <- data_adapt_rank(Y.param, A.param, W.param, absolute,
-    																			 negative)
-    return(data.adaptive.index)
+                                           negative)
+
+    index.grid <- which(data.adaptive.index <= n.top)
+    SL.lib <- c("SL.glm", "SL.step", "SL.glm.interaction", 'SL.gam', 'SL.earth')
+    psi_list <- list()
+    EIC_list <- list()
+    for(it_ind in seq_along(index.grid)){
+      print(index.grid[it_ind])
+      tmle.estimation <- tmle(Y = Y.est[, index.grid[it_ind]], A = A.est, W = W.est,
+                              Q.SL.library = SL.lib, g.SL.library = SL.lib)
+      psi_list[[it_ind]] <- tmle.estimation$estimates$ATE$psi
+      EIC_list[[it_ind]] <- tmle.estimation$estimates$IC$IC.ATE
+    }
+    psi.est <- do.call(c, psi_list)
+    EIC.est <- do.call(cbind, EIC_list)
+
+    return(list(data.adaptive.index, index.grid, psi.est, EIC.est))
   }
 
   # ============================================================================
   # CV
   # ============================================================================
-  if (parallel) {
-    doParallel::registerDoParallel(parallel::detectCores())
-    rank.all.fold <- foreach(it2 = 1:n.fold, .combine = cbind) %dopar% {
-      compute.a.fold(data_adapt, it2)
-    }
-  } else {
-    for (it0 in 1:n.fold) {
-      data.adaptive.index <- compute.a.fold(data_adapt, it0)
-      rank.all.fold[it0,] <- data.adaptive.index
-    }
+  for (it0 in 1:n.fold) {
+    list[data.adaptive.index, index.grid_here, psi.est_here, EIC.est_here] <- compute.a.fold(data_adapt, it0)
+    rank.all.fold[it0,] <- data.adaptive.index
+    adapt_param_composition[it0,] <- index.grid_here
+    psi.est_composition[[it0]] <- psi.est_here
+    EIC.est_composition[[it0]] <- EIC.est_here
   }
[…]
