+++ pkg/R/zzz.R
@@ -1,5 +1,5 @@
 .onAttach <- function(...) {
-  packageStartupMessage("data.adapt.multi.test: Data-adaptive test statistics for multiple testing in high-dimensional settings")
+  packageStartupMessage("DA.Test: Data-adaptive test statistics for multiple testing in high-dimensional settings")
   packageStartupMessage("Version: ",
-                        utils::packageDescription("data.adapt.multi.test")$Version)
+                        utils::packageDescription("DA.Test")$Version)
 }
