+++ pkg/man/get_composition.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/table.data_adapt.R
-\name{table.data_adapt}
-\alias{table.data_adapt}
+% Please edit documentation in R/get_composition.R
+\name{get_composition}
+\alias{get_composition}
 \title{Generic method of \code{table} for objects of class \code{data_adapt}}
 \usage{
-table.data_adapt(adapt.fit)
+get_composition(adapt.fit, type = "small")
 }
 \arguments{
 \item{adapt.fit}{data-adaptive statistical object of class \code{data_adapt}
