+++ pkg/man/get_significant_biomarker.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/get_composition.R
+\name{get_significant_biomarker}
+\alias{get_significant_biomarker}
+\title{Title}
+\usage{
+get_significant_biomarker(adapt.fit)
+}
+\arguments{
+\item{adapt.fit}{}
+}
+\description{
+Title
+}
+\examples{
+# NA
+}
