+++ pkg/NAMESPACE
@@ -6,9 +6,10 @@ S3method(print,data_adapt)
 export(data_adapt)
 export(data_adapt_multi_test)
 export(data_adapt_rank)
+export(get_composition)
 export(get_pval)
+export(get_significant_biomarker)
 export(shinyprint.data_adapt)
-export(table.data_adapt)
 importFrom(R2HTML,HTML)
 importFrom(calibrate,textxy)
 importFrom(doParallel,registerDoParallel)
