### 2017-10-05 e8caaff
remove plots in mcmc simu
