### 2017-10-05 085d442
edge case with no outptu
