+++ sim/paper_mcmc.R
@@ -51,17 +51,8 @@ simulate_once <- function(n.sim, p.all, p.true, signal.true, n.top.want, n.fold)
     gc()
     return(list(Y, A.sample.vec))
 }
-# ========================================================================================
-
-data_here <- simulate_once(n.sim, p.all, p.true, signal.true, n.top.want, n.fold)
-Y = data_here[[1]]
-A.sample.vec = data_here[[2]]
-
-
-SNR <- (signal.true^2 /4 + 1) / 1
-SNR
 # ==============================================================================
-# output
+# compute confusion matrix
 # ==============================================================================
 confusion_count <- function(vec_sig) {
     if(length(vec_sig) == 0){
@@ -77,20 +68,39 @@ confusion_count <- function(vec_sig) {
     }
     return(matrix(c(ONE,TWO,THREE,FOUR), nrow = 2, ncol = 2, byrow = TRUE))
 }
-# ==============================================================================
-# fit DA.test
-# ==============================================================================
-library(DA.Test)
-out_result <- data_adapt_multi_test(Y = Y, A = A.sample.vec, n.top = p.true + 5,
-                                    # n.fold = 3, parallel = FALSE) # BH on all Y
-                                    n.fold = 4, parallel = FALSE) # BH on all Y
-out_result
-col_id_sig_final <- as.numeric(DA.Test::get_significant_biomarker(out_result))
-confusion_count(col_id_sig_final)
+# ========================================================================================
+NUM_REPEAT = 1e1
+all_DA_Test_confuse <- list()
+all_naive_BH_confuse <- list()
+for (it in 1:NUM_REPEAT) {
+    print(it)
+    data_here <- simulate_once(n.sim, p.all, p.true, signal.true, n.top.want, n.fold)
+    Y = data_here[[1]]
+    A.sample.vec = data_here[[2]]
 
-# ==============================================================================
-# fit naive BH
-# ==============================================================================
-lm_out <- lm(Y ~ A.sample.vec)
-B1_hat_lm <- lm_out$coefficients[2,]
-names(B1_hat_lm) <- c(1:ncol(Y))
+    SNR <- (signal.true^2 /4 + 1) / 1
+    SNR
+    # fit DA.test
+    library(DA.Test)
+    out_result <- data_adapt_multi_test(Y = Y, A = A.sample.vec, n.top = p.true + 5,
+                                        # n.fold = 3, parallel = FALSE) # BH on all Y
+                                        n.fold = 4, parallel = FALSE) # BH on all Y
+    out_result
+    col_id_sig_final <- as.numeric(DA.Test::get_significant_biomarker(out_result))
+    DA_Test_confuse <- confusion_count(col_id_sig_final)
+
+    # fit naive BH
+    lm_out <- lm(Y ~ A.sample.vec)
+    B1_hat_lm <- lm_out$coefficients[2,]
+    names(B1_hat_lm) <- c(1:ncol(Y))
+
+    lm_summary <- summary(lm_out)
+    pval_lm <- sapply(lm_summary, function(x) x$coefficients[2,4])
+    qval_lm <- p.adjust(pval_lm, method = 'BH')
+    lm_sig <- which(qval_lm < .05)
+
+    naive_BH_confuse <- confusion_count(lm_sig)
+
+    all_DA_Test_confuse[[it]] <- DA_Test_confuse
+    all_naive_BH_confuse[[it]] <- naive_BH_confuse
+}
