### 2017-10-05 b920e49
complete mcmc simulation
