+++ sim/paper_mcmc.R
@@ -69,8 +69,11 @@ confusion_count <- function(vec_sig) {
     return(matrix(c(ONE,TWO,THREE,FOUR), nrow = 2, ncol = 2, byrow = TRUE))
 }
 # ========================================================================================
+set.seed(1234)
 NUM_REPEAT = 1e1
+# NUM_REPEAT = 5
 all_DA_Test_confuse <- list()
+all_wrong_algo_confuse <- list()
 all_naive_BH_confuse <- list()
 for (it in 1:NUM_REPEAT) {
     print(it)
@@ -81,13 +84,19 @@ for (it in 1:NUM_REPEAT) {
     SNR <- (signal.true^2 /4 + 1) / 1
     SNR
     # fit DA.test
-    library(DA.Test)
+    library(adaptest)
     out_result <- data_adapt_multi_test(Y = Y, A = A.sample.vec, n.top = p.true + 5,
-                                        # n.fold = 3, parallel = FALSE) # BH on all Y
                                         n.fold = 4, parallel = FALSE) # BH on all Y
     out_result
-    col_id_sig_final <- as.numeric(DA.Test::get_significant_biomarker(out_result))
+    col_id_sig_final <- as.numeric(adaptest::get_significant_biomarker(out_result))
     DA_Test_confuse <- confusion_count(col_id_sig_final)
+
+    # fit old algorithm
+    out_result_2 <- data.adapt.multi.test::data_adapt_multi_test(Y = Y, A = A.sample.vec, n.top = p.true + 5,
+                                        n.fold = 4, parallel = FALSE) # BH on all Y
+    out_result_2
+    col_id_sig_final <- out_result_2$sig.p.FDR
+    wrong_algo_confuse <- confusion_count(col_id_sig_final)
 
     # fit naive BH
     lm_out <- lm(Y ~ A.sample.vec)
@@ -102,5 +111,10 @@ for (it in 1:NUM_REPEAT) {
     naive_BH_confuse <- confusion_count(lm_sig)
 
     all_DA_Test_confuse[[it]] <- DA_Test_confuse
+    all_wrong_algo_confuse[[it]] <- wrong_algo_confuse
     all_naive_BH_confuse[[it]] <- naive_BH_confuse
 }
+
+Reduce('+', all_DA_Test_confuse)/length(all_DA_Test_confuse)
+Reduce('+', all_wrong_algo_confuse)/length(all_wrong_algo_confuse)
+Reduce('+', all_naive_BH_confuse)/length(all_naive_BH_confuse)
