+++ pkg/R/get_pval.R
@@ -10,14 +10,13 @@
 #' @examples
 #' #NA
 get_pval <- function(Psi_output, EIC.est_final, alpha=0.05) {
-    n.sim <- nrow(EIC.est_final)
-	var_by_col <- apply(EIC.est_final, 2, var)/n.sim
-	sd_by_col <- sqrt(var_by_col)
-	upper <- Psi_output + 1.96 * sd_by_col
-	lower <- Psi_output - 1.96 * sd_by_col
+  n.sim <- nrow(EIC.est_final)
+  var_by_col <- apply(EIC.est_final, 2, var)/n.sim
+  sd_by_col <- sqrt(var_by_col)
+  upper <- Psi_output + 1.96 * sd_by_col
+  lower <- Psi_output - 1.96 * sd_by_col
 
-	# pnorm(abs(Psi_output), mean = 0, sd = 1, lower.tail = FALSE) * 2
-	pval <- pnorm(abs(Psi_output/sd_by_col), mean = 0, sd = 1, lower.tail = FALSE) * 2
+  pval <- pnorm(abs(Psi_output/sd_by_col), mean = 0, sd = 1, lower.tail = FALSE) * 2
 
-	return(list(pval, upper, lower, sd_by_col))
+  return(list(pval, upper, lower, sd_by_col))
 }
