+++ pkg/R/data_adapt_multi_test.R
@@ -8,9 +8,9 @@
 #' @param Y continuous or binary outcome variable
 #' @param A binary treatment indicator: \code{1} = treatment, \code{0} = control
 #' @param W matrix containing baseline covariates
-#' @param n.top integer value for the number of candidate covariates to generate
+#' @param n_top integer value for the number of candidate covariates to generate
 #'        using the data-adaptive estimation algorithm
-#' @param n.fold integer number of folds to be used for cross-validation
+#' @param n_fold integer number of folds to be used for cross-validation
 #' @param negative boolean: \code{TRUE} = test for negative effect size,
 #'        \code{FALSE} = test for positive effect size
 #' @param absolute boolean: \code{TRUE} = test for absolute effect size. This
@@ -35,8 +35,8 @@
 data_adapt_multi_test <- function(Y,
                                   A,
                                   W = NULL,
-                                  n.top,
-                                  n.fold,
+                                  n_top,
+                                  n_fold,
                                   absolute = FALSE,
                                   negative = FALSE,
                                   parallel = FALSE) {
@@ -48,18 +48,18 @@ data_adapt_multi_test <- function(Y,
   }
 
   # use constructor function to instantiate "data_adapt" object
-  data_adapt <- data_adapt(Y, A, W = NULL, n.top, n.fold, absolute = FALSE,
+  data_adapt <- data_adapt(Y, A, W = NULL, n_top, n_fold, absolute = FALSE,
                            negative = FALSE, parallel = FALSE)
 
   # ============================================================================
   # preparation
   # ============================================================================
-  n.sim <- nrow(data_adapt$Y)
-  p.all <- ncol(data_adapt$Y)
+  n_sim <- nrow(data_adapt$Y)
+  p_all <- ncol(data_adapt$Y)
 
   # if there is no W input, use intercept as W
   if (is.null(data_adapt$W)) {
-    W <- as.matrix(rep(1, n.sim))
+    W <- as.matrix(rep(1, n_sim))
     data_adapt$W <- W
   }
 
@@ -67,116 +67,116 @@ data_adapt_multi_test <- function(Y,
   # create parameter generating sample
   # ============================================================================
   # determine number of samples per fold
-  sample.each.fold <- ceiling(n.sim / n.fold)
+  sample_each_fold <- ceiling(n_sim / n_fold)
 
   # random index
-  n.index.param.gen <- sample(head(rep(1:n.fold, each = sample.each.fold),
-                                   n = n.sim))
+  index_for_folds <- sample(head(rep(1:n_fold, each = sample_each_fold),
+                                   n = n_sim))
 
   # number of observations in each fold
-  folds.table <- table(n.index.param.gen)
+  table_n_per_fold <- table(index_for_folds)
 
-  rank.all.fold <- matrix(0, nrow = n.fold, ncol = p.all)
-  adapt_param_composition <- matrix(0, nrow = n.fold, ncol = n.top)
-  psi.est_composition <- list()
-  EIC.est_composition <- list()
+  rank_in_folds <- matrix(0, nrow = n_fold, ncol = p_all)
+  adapt_param_composition <- matrix(0, nrow = n_fold, ncol = n_top)
+  psi_est_composition <- list()
+  EIC_est_composition <- list()
   # ============================================================================
[…]
