+++ pkg/R/plot.data_adapt.R
@@ -16,35 +16,35 @@
 #'
 plot.data_adapt <- function(x, ..., plot_type = c("cvrank", "pvals")) {
 
-  top.index <- x$top.index
-  ATE.subset <- x$ATE.subset
-  p.init <- x$p.init
-  p.final <- x$p.final
-  still.sig.p <- x$sig.p.FDR
+  top_index <- x$top_index
+  DE <- x$DE
+  p_value <- x$p_value
+  q_value <- x$q_value
+  significant_q <- x$significant_q
 
-  top.mean.rank <- x$top.mean.rank
-  p.in.top.rank <- x$p.in.top.rank
-  n.top.want <- length(top.index)
+  mean_rank_top <- x$mean_rank_top
+  prob_in_top <- x$prob_in_top
+  n_top.want <- length(top_index)
 
   if ("cvrank" %in% plot_type) {
     # Plot sorted average CV-rank
-    plot(top.mean.rank, ylab = 'Mean CV-rank', pch = 20,
+    plot(mean_rank_top, ylab = 'Mean CV-rank', pch = 20,
          main = 'Mean CV-rank of selected covariates \n (Smaller the better)')
 
-    calibrate::textxy((1:n.top.want) - 0.3, top.mean.rank + 0.5, top.index,
+    calibrate::textxy((1:n_top.want) - 0.3, mean_rank_top + 0.5, top_index,
                       offset = .6)
     abline(a = 0, b = 1, lty = 3)
   }
 
   if ("pvals" %in% plot_type) {
     # plot sorted p-values, labeled with index
-    temp.top.index <- c(1:n.top.want)[order(p.final)]
-    plot(sort(p.final),
+    temp.top_index <- c(1:n_top.want)[order(q_value)]
+    plot(sort(q_value),
          pch = 20,
          ylab = 'Adjusted P-value',
          main = 'Adjusted p-value of selected covariates \n (Smaller the better)'
          )
-    calibrate::textxy((1:n.top.want) - 0.3, sort(p.final), temp.top.index,
+    calibrate::textxy((1:n_top.want) - 0.3, sort(q_value), temp.top_index,
                       offset = 1)
     abline(h = 0.05, lty = 2)
   }
