+++ pkg/R/data_adapt_rank.R
@@ -4,10 +4,10 @@
 #' is designed to be called inside \code{data_adapt_multi_test}; it should not
 #' be run by itself outside of that contex.
 #'
-#' @param Y.param continuous or binary outcome variable
-#' @param A.param binary treatment indicator: \code{1} = treatment,
+#' @param Y_param continuous or binary outcome variable
+#' @param A_param binary treatment indicator: \code{1} = treatment,
 #'        \code{0} = control
-#' @param W.param vector, matrix, or data.frame containing baseline covariates
+#' @param W_param vector, matrix, or data.frame containing baseline covariates
 #' @param absolute boolean: \code{TRUE} = test for absolute effect size. This
 #'        \code{FALSE} = test for directional effect. This overrides argument
 #'        \code{negative}.
@@ -19,50 +19,49 @@
 #'
 #' @export data_adapt_rank
 #'
-data_adapt_rank <- function(Y.param,
-                            A.param,
-                            W.param,
+data_adapt_rank <- function(Y_param,
+                            A_param,
+                            W_param,
                             absolute = FALSE,
                             negative = FALSE) {
-  n.here <- nrow(Y.param)
-  p.all <- ncol(Y.param)
+  n_here <- nrow(Y_param)
+  p_all <- ncol(Y_param)
 
-  B1.fitted.all <- rep(0, p.all)
+  B1_fitted <- rep(0, p_all)
 
-  SL.lib <- c("SL.glm", "SL.step", "SL.glm.interaction", 'SL.gam')
+  SL_lib <- c("SL.glm", "SL.step", "SL.glm.interaction", 'SL.gam')
 
-  for (it in 1:p.all) {
-    A.fit <- A.param
-    Y.fit <- Y.param[,it]
-    W.fit <- as.matrix(W.param)
+  for (it in 1:p_all) {
+    A_fit <- A_param
+    Y_fit <- Y_param[,it]
+    W_fit <- as.matrix(W_param)
     # CASE 1: TMLE for effect size
-    if ( !identical(W.param, as.matrix(rep(1, n.here))) ) {
+    if ( !identical(W_param, as.matrix(rep(1, n_here))) ) {
       # if there are W
-      tmle.result <- tmle(Y = Y.fit, A = A.fit, W = W.param,
-                          Q.SL.library = SL.lib, g.SL.library = SL.lib)
-      B1.result <- tmle.result$estimates$ATE$psi
+      tmle_fit <- tmle(Y = Y_fit, A = A_fit, W = W_param,
+                          Q.SL_library = SL_lib, g.SL_library = SL_lib)
+      B1_fitted[it] <- tmle_fit$estimates$ATE$psi
     } else {
       # CASE 2: OLS for faster effect size
-      lm.result <- lm(Y.fit ~ A.fit)
-      B1.result <- lm.result$coefficients[2]
+      lm_fit <- lm(Y_fit ~ A_fit)
+      B1_fitted[it] <- lm_fit$coefficients[2]
     }
-    B1.fitted.all[it] <- B1.result
   }
 
   # rank by absolute differential expression
   if (absolute == TRUE) {
-    B1.fitted.all.abs <- abs(B1.fitted.all)
+    B1_fitted_abs <- abs(B1_fitted)
   } else {
-    B1.fitted.all.abs <- B1.fitted.all
+    B1_fitted_abs <- B1_fitted
   }
 
   # calculate rank of each covariate
   if (negative) {
-    rank.out <- rank(B1.fitted.all.abs)
+    rank_out <- rank(B1_fitted_abs)
   } else {
-    rank.out <- rank(-B1.fitted.all.abs)
+    rank_out <- rank(-B1_fitted_abs)
   }
 
   # final object to be exported by this function
-  return(rank.out)
+  return(rank_out)
 }
