### 2017-10-05 831fea4
change main function name
