+++ pkg/R/shinyprint.data_adapt.R
@@ -2,17 +2,17 @@
 #'
 #' Provides HTML-based printing utility for examining data-adaptive statistics
 #'
-#' @param adapt.fit data-adaptive statistical object of class \code{data_adapt}
-#'        as returned by \code{data_adapt_multi_test}
+#' @param object data-adaptive statistical object of class \code{data_adapt}
+#'        as returned by \code{adaptest}
 #'
 #' @importFrom R2HTML HTML
 #'
 #' @export shinyprint.data_adapt
 #'
-shinyprint.data_adapt <- function(adapt.fit) {
-  print.data_adapt(adapt.fit)
+shinyprint.data_adapt <- function(object) {
+  print.data_adapt(object)
   HTML(paste('<b> The covariates still significant are </b>',
-             paste(adapt.fit[[5]], collapse = ' '),
+             paste(object[[5]], collapse = ' '),
              sep = '<br/>'
   ))
 }
