+++ sim/paper_simulation.Rmd
@@ -61,7 +61,7 @@ gc()
 # fit DA.test
 
 ```{r}
-library(data.adapt.multi.test)
+library(adaptest)
 out_result <- data_adapt_multi_test(Y = Y, A = A.sample.vec, n.top = p.true + 5,
                                     n.fold = 4, parallel = FALSE) # BH on all Y
 print(out_result)
