+++ pkg/man/plot.data_adapt.Rd
@@ -8,7 +8,7 @@
 }
 \arguments{
 \item{x}{data-adaptive statistical object of class \code{data_adapt} as
-returned by \code{data_adapt_multi_test}}
+returned by \code{adaptest}}
 
 \item{...}{additional arguments passed to \code{plot} as necessary}
 
