+++ pkg/man/adaptest.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/data_adapt_multi_test.R
-\name{data_adapt_multi_test}
-\alias{data_adapt_multi_test}
+% Please edit documentation in R/adaptest.R
+\name{adaptest}
+\alias{adaptest}
 \title{Data-adaptive test statistics for high-dimensional multiple testing}
 \usage{
-data_adapt_multi_test(Y, A, W = NULL, n_top, n_fold, absolute = FALSE,
+adaptest(Y, A, W = NULL, n_top, n_fold, absolute = FALSE,
   negative = FALSE, parallel = FALSE)
 }
 \arguments{
