### 2017-10-06 2a9c0d6
combine multiple function into one file
