+++ pkg/man/get_significant_biomarker.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/get_composition.R
+% Please edit documentation in R/get_summary.R
 \name{get_significant_biomarker}
 \alias{get_significant_biomarker}
 \title{Title}
