+++ pkg/man/get_composition.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/get_composition.R
+% Please edit documentation in R/get_summary.R
 \name{get_composition}
 \alias{get_composition}
 \title{Generic method of \code{table} for objects of class \code{data_adapt}}
