+++ pkg/man/rank_DE.Rd
@@ -1,19 +1,18 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/data_adapt_rank.R
-\name{data_adapt_rank}
-\alias{data_adapt_rank}
+% Please edit documentation in R/generate_parameter_wrapper.R
+\name{rank_DE}
+\alias{rank_DE}
 \title{Ranking for data-adaptive test statistics}
 \usage{
-data_adapt_rank(Y_param, A_param, W_param, absolute = FALSE,
-  negative = FALSE)
+rank_DE(Y, A, W, absolute = FALSE, negative = FALSE)
 }
 \arguments{
-\item{Y_param}{continuous or binary outcome variable}
+\item{Y}{continuous or binary outcome variable}
 
-\item{A_param}{binary treatment indicator: \code{1} = treatment,
+\item{A}{binary treatment indicator: \code{1} = treatment,
 \code{0} = control}
 
-\item{W_param}{vector, matrix, or data.frame containing baseline covariates}
+\item{W}{vector, matrix, or data.frame containing baseline covariates}
 
 \item{absolute}{boolean: \code{TRUE} = test for absolute effect size. This
 \code{FALSE} = test for directional effect. This overrides argument
