+++ pkg/man/plot.data_adapt.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/print_plot_shinyprint.R
+% Please edit documentation in R/print_plot_shinyprint_table.R
 \name{plot.data_adapt}
 \alias{plot.data_adapt}
 \title{Generic plot method for data_adapt class}
