+++ pkg/man/print.data_adapt.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/print_plot_shinyprint.R
+% Please edit documentation in R/print_plot_shinyprint_table.R
 \name{print.data_adapt}
 \alias{print.data_adapt}
 \title{Generic print method for data_adapt class}
