+++ pkg/man/summary.data_adapt.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/print_plot_shinyprint_tabular.R
+\name{summary.data_adapt}
+\alias{summary.data_adapt}
+\title{Title}
+\usage{
+\method{summary}{data_adapt}(x, ..., type = "adapt_param")
+}
+\arguments{
+\item{type}{}
+}
+\description{
+Title
+}
+\examples{
+#NA
+}
