+++ pkg/man/plot.data_adapt.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/print_plot_shinyprint_table.R
+% Please edit documentation in R/print_plot_shinyprint_tabular.R
 \name{plot.data_adapt}
 \alias{plot.data_adapt}
 \title{Generic plot method for data_adapt class}
 \usage{
-\method{plot}{data_adapt}(x, ..., plot_type = c("cvrank", "pvals"))
+\method{plot}{data_adapt}(x, ..., plot_type = c("biomarker", "adapt_param"))
 }
 \arguments{
 \item{x}{data-adaptive statistical object of class \code{data_adapt} as
@@ -13,7 +13,7 @@ returned by \code{adaptest}}
 \item{...}{additional arguments passed to \code{plot} as necessary}
 
 \item{plot_type}{character vector specifying which of the two types of plots
-to generate: "cvrank" for a plot sorted average CV-rank, or "pvals"
+to generate: "biomarker" for a plot sorted average CV-rank, or "adapt_param"
 for a plot sorted by p-values with labels corresponding to indices}
 }
 \description{
