+++ pkg/NAMESPACE
@@ -3,6 +3,7 @@
 S3method("[<-",result)
 S3method(plot,data_adapt)
 S3method(print,data_adapt)
+S3method(summary,data_adapt)
 export(adaptest)
 export(data_adapt)
 export(get_composition)
