### 2017-10-16 f7f54aa
clean simu code
