+++ pkg/man/shinyprint.data_adapt.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/print_plot_shinyprint_tabular.R
+% Please edit documentation in R/print_plot_shinyprint_summary.R
 \name{shinyprint.data_adapt}
 \alias{shinyprint.data_adapt}
 \title{Method of \code{shinyprint} for objects of class \code{data_adapt}}
