+++ pkg/man/summary.data_adapt.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/print_plot_shinyprint_tabular.R
+% Please edit documentation in R/print_plot_shinyprint_summary.R
 \name{summary.data_adapt}
 \alias{summary.data_adapt}
 \title{Title}
