### 2017-10-16 eec1064
remove useless SL_lib
