+++ R/adaptest.R
@@ -1,4 +1,4 @@
-#' Constructor function for class \code{data_adapt}
+#' Constructor function for class data_adapt
 #'
 #' @param Y continuous or binary outcome variable
 #' @param A binary treatment indicator: \code{1} = treatment, \code{0} = control
@@ -17,7 +17,7 @@
 #' @return S3 object of class "data_adapt" for data-adaptive multiple testing.
 #'
 #' @export data_adapt
-#'
+#
 data_adapt <- function(Y,
                        A,
                        W = NULL,
@@ -41,7 +41,9 @@ data_adapt <- function(Y,
   if (!is.numeric(n_fold)) stop("argument n_fold must be numeric")
   if (!is.logical(absolute)) stop("argument absolute must be boolean/logical")
   if (!is.logical(negative)) stop("argument negative must be boolean/logical")
-  if (!is.function(parameter_wrapper)) stop("argument parameter_wrapper must be function")
+  if (!is.function(parameter_wrapper)) {
+    stop("argument parameter_wrapper must be function")
+  }
   if (!is.character(SL_lib)) stop("argument SL_lib must be character")
 
   # placeholders for outputs to be included when returning the data_adapt object
@@ -53,20 +55,23 @@ data_adapt <- function(Y,
   mean_rank_top <- NULL
   prob_in_top <- NULL
 
-  out <- structure(list(Y, A, W, n_top, n_fold, absolute, negative, parameter_wrapper, SL_lib,
+  out <- structure(list(Y, A, W, n_top, n_fold, absolute, negative,
+                        parameter_wrapper, SL_lib,
                         top_colname, DE, p_value, q_value, significant_q,
                         mean_rank_top, prob_in_top),
                    class = "data_adapt")
 
   names(out) <- c("Y", "A", "W", "n_top", "n_fold", "absolute", "negative",
-                  "parameter_wrapper", "SL_lib", "top_colname", "DE", "p_value", "q_value",
-                  "significant_q", "mean_rank_top", "prob_in_top")
+                  "parameter_wrapper", "SL_lib", "top_colname", "DE", "p_value",
+                  "q_value", "significant_q", "mean_rank_top", "prob_in_top")
 
   # export instance of "data_adapt" for downstream use
   return(out)
 }
 
-#' Title
+################################################################################
+
+#' Extract p-values
 #'
 #' @param Psi_output
 #' @param EIC_est_final
@@ -74,9 +79,7 @@ data_adapt <- function(Y,
 #'
 #' @return
 #' @export
-#'
-#' @examples
-#' #NA
+#
 get_pval <- function(Psi_output, EIC_est_final, alpha=0.05) {
   n_sim <- nrow(EIC_est_final)
   var_by_col <- apply(EIC_est_final, 2, var)/n_sim
@@ -89,7 +92,7 @@ get_pval <- function(Psi_output, EIC_est_final, alpha=0.05) {
   return(list(pval, upper, lower, sd_by_col))
 }
 
-
+################################################################################
 
 #' Data-adaptive test statistics for high-dimensional multiple testing
 #'
@@ -102,38 +105,37 @@ get_pval <- function(Psi_output, EIC_est_final, alpha=0.05) {
 #' @param A binary treatment indicator: \code{1} = treatment, \code{0} = control
 #' @param W matrix containing baseline covariates
 #' @param n_top integer value for the number of candidate covariates to generate
-#'        using the data-adaptive estimation algorithm
+#'  using the data-adaptive estimation algorithm
 #' @param n_fold integer number of folds to be used for cross-validation
 #' @param negative boolean: \code{TRUE} = test for negative effect size,
-#'        \code{FALSE} = test for positive effect size
+#'  \code{FALSE} = test for positive effect size
[…]
