### 2017-10-19 404afc0
add origami+future test
