### 2017-10-22 463c341
remove slow W code
