### 2017-10-24 62d3550
fix set.seed error in test
