### 2017-10-24 c12afcf
cv_fun changed
