+++ R/adaptest_old.R
@@ -1,0 +1,183 @@
+#' OLDOLDOLDO
+#'
+#' Performs targeted minimum loss-based estimation (TMLE )of a marginal additive
+#' treatment effect of a binary point treatment on an outcome. The data-adaptive
+#' algorithm is used to perform variable reduction to avoid the disadvantages
+#' associated with multiple testing.
+#'
+#' @param Y continuous or binary outcome variable
+#' @param A binary treatment indicator: \code{1} = treatment, \code{0} = control
+#' @param W matrix containing baseline covariates
+#' @param n_top integer value for the number of candidate covariates to generate
+#'        using the data-adaptive estimation algorithm
+#' @param n_fold integer number of folds to be used for cross-validation
+#' @param negative boolean: \code{TRUE} = test for negative effect size,
+#'        \code{FALSE} = test for positive effect size
+#' @param absolute boolean: \code{TRUE} = test for absolute effect size. This
+#'        \code{FALSE} = test for directional effect. This overrides argument
+#'        \code{negative}.
+#' @param parameter_wrapper function
+#' @param SL_lib character
+#'
+#' @importFrom tmle tmle
+#' @importFrom foreach foreach "%dopar%"
+#' @importFrom parallel detectCores
+#' @importFrom doParallel registerDoParallel
+#' @importFrom stats lm p.adjust
+#' @importFrom utils head
+#' @importFrom magrittr "%>%"
+#'
+#' @author Wilson Cai \email{wcai@@berkeley.edu}, in collaboration with Alan E.
+#'         Hubbard, with contributions from Nima S. Hejazi.
+#'
+#'
+adaptest_old <- function(Y,
+                     A,
+                     W = NULL,
+                     n_top,
+                     n_fold,
+                     parameter_wrapper = adaptest::rank_DE,
+                     SL_lib = c("SL.glm", "SL.step", "SL.glm.interaction", 'SL.gam', 'SL.earth'),
+                     absolute = FALSE,
+                     negative = FALSE) {
+  # use constructor function to instantiate "data_adapt" object
+  data_adapt <- data_adapt(Y = Y, A = A, W = W,
+                           n_top = n_top,
+                           n_fold = n_fold,
+                           absolute = absolute,
+                           negative = negative,
+                           parameter_wrapper = parameter_wrapper,
+                           SL_lib = SL_lib)
+  # ============================================================================
+  # preparation
+  # ============================================================================
+  n_sim <- nrow(data_adapt$Y)
+  p_all <- ncol(data_adapt$Y)
+
+  # if there is no W input, use intercept as W
+  if (is.null(data_adapt$W)) {
+    W <- as.matrix(rep(1, n_sim))
+    data_adapt$W <- W
+  }
+
+  # ============================================================================
+  # create parameter generating sample
+  # ============================================================================
+  # determine number of samples per fold
+  sample_each_fold <- ceiling(n_sim / n_fold)
+
+  # random index
+  index_for_folds <- sample(head(rep(1:n_fold, each = sample_each_fold),
+                                   n = n_sim))
+
+  # number of observations in each fold
+  table_n_per_fold <- table(index_for_folds)
+
+  rank_in_folds <- matrix(0, nrow = n_fold, ncol = p_all)
+  adapt_param_composition <- matrix(0, nrow = n_fold, ncol = n_top)
+  psi_est_composition <- list()
+  EIC_est_composition <- list()
+  # ============================================================================
+  compute.a.fold <- function(data_adapt, it0) {
+    print(paste('Fold:', it0))
+    chunk_as_est <- it0
+
[…]
