+++ man/adaptest_old.Rd
@@ -1,0 +1,44 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/adaptest_old.R
+\name{adaptest_old}
+\alias{adaptest_old}
+\title{OLDOLDOLDO}
+\usage{
+adaptest_old(Y, A, W = NULL, n_top, n_fold,
+  parameter_wrapper = adaptest::rank_DE, SL_lib = c("SL.glm", "SL.step",
+  "SL.glm.interaction", "SL.gam", "SL.earth"), absolute = FALSE,
+  negative = FALSE)
+}
+\arguments{
+\item{Y}{continuous or binary outcome variable}
+
+\item{A}{binary treatment indicator: \code{1} = treatment, \code{0} = control}
+
+\item{W}{matrix containing baseline covariates}
+
+\item{n_top}{integer value for the number of candidate covariates to generate
+using the data-adaptive estimation algorithm}
+
+\item{n_fold}{integer number of folds to be used for cross-validation}
+
+\item{parameter_wrapper}{function}
+
+\item{SL_lib}{character}
+
+\item{absolute}{boolean: \code{TRUE} = test for absolute effect size. This
+\code{FALSE} = test for directional effect. This overrides argument
+\code{negative}.}
+
+\item{negative}{boolean: \code{TRUE} = test for negative effect size,
+\code{FALSE} = test for positive effect size}
+}
+\description{
+Performs targeted minimum loss-based estimation (TMLE )of a marginal additive
+treatment effect of a binary point treatment on an outcome. The data-adaptive
+algorithm is used to perform variable reduction to avoid the disadvantages
+associated with multiple testing.
+}
+\author{
+Wilson Cai \email{wcai@berkeley.edu}, in collaboration with Alan E.
+        Hubbard, with contributions from Nima S. Hejazi.
+}
