+++ NAMESPACE
@@ -13,6 +13,9 @@ export(rank_DE)
 export(shinyprint.data_adapt)
 importFrom(R2HTML,HTML)
 importFrom(calibrate,textxy)
+importFrom(doParallel,registerDoParallel)
+importFrom(foreach,"%dopar%")
+importFrom(foreach,foreach)
 importFrom(graphics,abline)
 importFrom(graphics,plot)
 importFrom(magrittr,"%>%")
@@ -20,6 +23,7 @@ importFrom(origami,cross_validate)
 importFrom(origami,make_folds)
 importFrom(origami,training)
 importFrom(origami,validation)
+importFrom(parallel,detectCores)
 importFrom(stats,lm)
 importFrom(stats,p.adjust)
 importFrom(tmle,tmle)
