### 2017-10-24 c797a85
minor test changes
