+++ R/adaptest.R
@@ -173,7 +173,7 @@ adaptest <- function(Y,
 
   # ============================================================================
   folds <- origami::make_folds(n = n_sim, V = n_fold)
-  df_all <- data.frame(Y = Y, A = A.sample.vec, W = W)
+  df_all <- data.frame(Y = Y, A = A, W = W)
   cv_results <- origami::cross_validate(cv_fun = cv_param_est, folds = folds,
                                         data = df_all,
                                         parameter_wrapper = parameter_wrapper,
