### 2017-10-24 be21b5e
add deps, update old-new test
