+++ R/zzz.R
@@ -1,5 +1,7 @@
 .onAttach <- function(...) {
-  packageStartupMessage("adaptest: Data-adaptive test statistics for multiple testing in high-dimensional settings")
+  packageStartupMessage(paste("adaptest: Data-adaptive statistics for",
+                              "\n high-dimensional multiple testing"))
   packageStartupMessage("Version: ",
                         utils::packageDescription("adaptest")$Version)
 }
+
