+++ R/adaptest.R
@@ -1,4 +1,4 @@
-#' Constructor function for class data_adapt
+#' Class Constructor for Data Adaptive Parameters
 #'
 #' @param Y continuous or binary outcome variable
 #' @param A binary treatment indicator: \code{1} = treatment, \code{0} = control
@@ -15,8 +15,6 @@
 #' @param SL_lib character
 #'
 #' @return S3 object of class "data_adapt" for data-adaptive multiple testing.
-#'
-#' @export data_adapt
 #
 data_adapt <- function(Y,
                        A,
@@ -54,16 +52,18 @@ data_adapt <- function(Y,
   significant_q <- NULL
   mean_rank_top <- NULL
   prob_in_top <- NULL
+  folds <- NULL
 
   out <- structure(list(Y, A, W, n_top, n_fold, absolute, negative,
-                        parameter_wrapper, SL_lib,
-                        top_colname, DE, p_value, q_value, significant_q,
-                        mean_rank_top, prob_in_top),
+                        parameter_wrapper, SL_lib, top_colname, DE,
+                        p_value, q_value, significant_q,
+                        mean_rank_top, prob_in_top, folds),
                    class = "data_adapt")
 
   names(out) <- c("Y", "A", "W", "n_top", "n_fold", "absolute", "negative",
                   "parameter_wrapper", "SL_lib", "top_colname", "DE", "p_value",
-                  "q_value", "significant_q", "mean_rank_top", "prob_in_top")
+                  "q_value", "significant_q", "mean_rank_top", "prob_in_top",
+                  "folds")
 
   # export instance of "data_adapt" for downstream use
   return(out)
@@ -71,7 +71,7 @@ data_adapt <- function(Y,
 
 ################################################################################
 
-#' Extract p-values
+#' Statistical Inference for Data-Adaptive Parameters
 #'
 #' @param Psi_output
 #' @param EIC_est_final
@@ -80,21 +80,22 @@ data_adapt <- function(Y,
 #' @return
 #' @export
 #
-get_pval <- function(Psi_output, EIC_est_final, alpha=0.05) {
+get_pval <- function(Psi_output, EIC_est_final, alpha = 0.05) {
   n_sim <- nrow(EIC_est_final)
-  var_by_col <- apply(EIC_est_final, 2, var)/n_sim
+  var_by_col <- apply(EIC_est_final, 2, var) / n_sim
   sd_by_col <- sqrt(var_by_col)
-  upper <- Psi_output + 1.96 * sd_by_col
-  lower <- Psi_output - 1.96 * sd_by_col
+  upper <- Psi_output + abs(qnorm(alpha / 2)) * sd_by_col
+  lower <- Psi_output - abs(qnorm(alpha / 2)) * sd_by_col
 
-  pval <- pnorm(abs(Psi_output/sd_by_col), mean = 0, sd = 1, lower.tail = FALSE) * 2
+  pval <- pnorm(abs(Psi_output / sd_by_col), mean = 0, sd = 1,
+                lower.tail = FALSE) * 2
 
   return(list(pval, upper, lower, sd_by_col))
 }
 
 ################################################################################
 
-#' Data-adaptive test statistics for high-dimensional multiple testing
+#' Data-adaptive Statistics for High-Dimensional Multiple Testing
 #'
 #' Performs targeted minimum loss-based estimation (TMLE )of a marginal additive
 #' treatment effect of a binary point treatment on an outcome. The data-adaptive
@@ -120,8 +121,7 @@ get_pval <- function(Psi_output, EIC_est_final, alpha=0.05) {
 #' @importFrom magrittr "%>%"
[…]
