+++ tests/testthat/test-old_vs_new.R
@@ -6,6 +6,7 @@ library(MASS)
 library(Matrix)
 library(future)
 library(adaptest)
+library(origami)
 context("adaptest works the same under old and origami implementations")
 
 ################################################################################
@@ -20,7 +21,7 @@ p.all <- 1e4 # No. of dimensions of Y (signal + non-signal)
 # p.all <- 1e6 # No. of dimensions of Y (signal + non-signal)
 p.true <- 10 # No. of dimensions for signal
 # signal.true <- 1 # size of true signal
-signal.true <- .6 # size of true signal
+signal.true <- 0.6 # size of true signal
 # signal.true <- .8 # size of true signal
 signal.true.W <- 0.1
 
@@ -35,12 +36,12 @@ epsilon <- matrix(rnorm(n = n.sim * p.all), nrow = n.sim, ncol = p.all)
 A.candidate <- list(rep(1, p.all), rep(0, p.all))
 A.sample <- sample(A.candidate, size = n.sim, replace = TRUE)
 A.sample <- do.call(rbind, A.sample)
-A.sample.vec <- A.sample[,1]
+A.sample.vec <- A.sample[, 1]
 
 # B1
 b1.row <- c(rep(signal.true, p.true), rep(0, p.all - p.true))
-rep.row<-function(x,n){
-  matrix(rep(x,each=n),nrow=n)
+rep.row <- function(x, n) {
+  matrix(rep(x, each = n),nrow = n)
 }
 b1 <- rep.row(b1.row, n = n.sim)
 
@@ -53,30 +54,83 @@ temp1 <- b1 * A.sample
 Y <- b0 + temp1 + epsilon
 
 # clean up
-rm(list = c('b1', 'A.sample'))
-rm(list = c('b0', 'epsilon', 'temp1'))
+rm(list = c("b1", "A.sample"))
+rm(list = c("b0", "epsilon", "temp1"))
 
 
 ################################################################################
 # adaptest should be fast with futures...
 ################################################################################
-plan(multiprocess)
+
+if(availableCores() > 1) {
+  plan(multiprocess)
+}
+
 set.seed(48915672)
 time_new <- system.time(
-    result_new <- adaptest(Y = Y, A = A.sample.vec, n_top = p.true + 5,
+    result_new <- adaptest(Y = Y,
+                           A = A.sample.vec,
+                           n_top = p.true + 5,
                            n_fold = 4)
 )
 
-set.seed(48915672)
+# get origami folds object generated by adaptest to feed into old implementaiton
+folds <- result_new$folds
+folds_vec <- origami::folds2foldvec(folds)
+
 time_old <- system.time(
-    result_old <- adaptest:::adaptest_old(Y = Y, A = A.sample.vec,
-                                          n_top = p.true + 5, n_fold = 4)
+    result_old <- adaptest:::adaptest_old(Y = Y,
+                                          A = A.sample.vec,
+                                          n_top = p.true + 5,
+                                          n_fold = 4,
+                                          folds_vec = folds_vec)
 )
 
-test_that("New and old routines return same data-adaptive target parameter", {
+# test individual slots in the data_adapt output objects for equivalence
+# NOTE: expect_equal() invoked directly on the two objects will return a
+#       non-equivalence since the new implementation appends an extra slot
+#       "folds" to the output object (needed to make these tests possible)
+test_that("New and old routines return same top mean ranks", {
   expect_equal(result_new$mean_rank_top, result_old$mean_rank_top)
 })
-test_that("New and old routines return identical objects", {
-  expect_equal(result_new, result_old)
+
+test_that("New and old routines return same top index", {
+  expect_equal(result_new$top_index, result_old$top_index)
 })
+
+test_that("New and old routines return same top column names", {
[…]
