+++ README.Rmd
@@ -1,5 +1,5 @@
 ---
-output: github_document
+output: rmarkdown::github_document
 bibliography: README-refs.bib
 ---
 
