+++ Makefile
@@ -1,5 +1,5 @@
 md:
-	Rscript -e "rmarkdown::render('README.Rmd')"
+	Rscript -e "rmarkdown::render('README.Rmd', output_file = 'README.md')"
 
 site:
 	Rscript -e "pkgdown::build_site()"
@@ -16,5 +16,9 @@ test:
 doc:
 	Rscript -e "devtools::document()"
 
+build:
+	Rscript -e "devtools::build()"
+
 cov:
 	Rscript -e "covr::package_coverage(type = 'all', combine_types = FALSE, line_exclusions = list('R/plots.R'))"
+

[…]
