+++ NAMESPACE
@@ -5,7 +5,6 @@ S3method(plot,data_adapt)
 S3method(print,data_adapt)
 S3method(summary,data_adapt)
 export(adaptest)
-export(data_adapt)
 export(get_composition)
 export(get_pval)
 export(get_significant_biomarker)
@@ -22,5 +21,8 @@ importFrom(origami,training)
 importFrom(origami,validation)
 importFrom(stats,lm)
 importFrom(stats,p.adjust)
+importFrom(stats,pnorm)
+importFrom(stats,qnorm)
+importFrom(stats,var)
 importFrom(tmle,tmle)
 importFrom(utils,head)
