### 2017-11-17 5820132
fix s3 method
