+++ R/utils.R
@@ -112,15 +112,15 @@ plot.data_adapt <- function(x, ..., plot_type = c("biomarker", "adapt_param")) {
 #'
 #' @examples
 #' #NA
-summary.data_adapt <- function(x, ..., type = 'adapt_param') {
-  data_adapt_param = 1:length(x$DE)
-  DE = x$DE
-  p_value = x$p_value
-  q_value = x$q_value
+summary.data_adapt <- function(object, ..., type = 'adapt_param') {
+  data_adapt_param = 1:length(object$DE)
+  DE = object$DE
+  p_value = object$p_value
+  q_value = object$q_value
 
-  top_biomarker = x$top_index
-  mean_rank_top = x$mean_rank_top
-  prob_in_top = x$prob_in_top*100
+  top_biomarker = object$top_index
+  mean_rank_top = object$mean_rank_top
+  prob_in_top = object$prob_in_top*100
 
   if (type == 'adapt_param'){
     table_out <- data.frame(data_adapt_param, DE, p_value, q_value)
