### 2017-11-17 143ece6
remove helper. include all options into roxygen
