### 2017-11-17 bb9918b
more roxygen arguments
