+++ NAMESPACE
@@ -1,6 +1,5 @@
 # Generated by roxygen2: do not edit by hand
 
-S3method("[<-",result)
 S3method(plot,data_adapt)
 S3method(print,data_adapt)
 S3method(summary,data_adapt)
