### 2017-11-17 e692adc
clean some code
