### 2017-11-17 55b4242
include more detailed return values
