### 2017-11-17 b242c0b
all warnings fixed
