### 2017-11-20 7b49377
fixes dependencies,coverage,etc -- tests still broken
