+++ tests/testthat/test-data_adapt.R
@@ -28,3 +28,4 @@ test_that("data_adapt fails when invoked improperly", {
 #               )
 
 # })
+
