+++ man/get_significant_biomarker.Rd
@@ -2,7 +2,7 @@
 % Please edit documentation in R/get_summary.R
 \name{get_significant_biomarker}
 \alias{get_significant_biomarker}
-\title{Title}
+\title{Extract statisticall significant biomarkers}
 \usage{
 get_significant_biomarker(object)
 }
@@ -10,8 +10,5 @@ get_significant_biomarker(object)
 \item{object}{\code{data_adapt} object}
 }
 \description{
-Title
-}
-\examples{
-# NA
+Extract statisticall significant biomarkers
 }
