+++ man/data_adapt.Rd
@@ -2,7 +2,7 @@
 % Please edit documentation in R/adaptest.R
 \name{data_adapt}
 \alias{data_adapt}
-\title{Class Constructor for Data Adaptive Parameters}
+\title{S3-Style Constructor for Data Adaptive Parameter Class}
 \usage{
 data_adapt(Y, A, W = NULL, n_top, n_fold, absolute, negative,
   parameter_wrapper, SL_lib)
@@ -31,8 +31,9 @@ using the data-adaptive estimation algorithm}
 \item{SL_lib}{character of SuperLearner library}
 }
 \value{
-\code{S3} object of class "data_adapt" for data-adaptive multiple testing.
+\code{S3} object of class "data_adapt" for data-adaptive multiple
+ testing.
 }
 \description{
-Class Constructor for Data Adaptive Parameters
+S3-Style Constructor for Data Adaptive Parameter Class
 }

[…]
