### 2017-11-21 03c9e42
rm bad test; fix multiprocess issue for appveyor
