### 2017-11-21 0a80714
rm earth from SL libs in tests
