+++ tests/testthat/test-old_vs_new.R
@@ -73,19 +73,23 @@ time_new <- system.time(
     result_new <- adaptest(Y = Y,
                            A = A.sample.vec,
                            n_top = p.true + 5,
-                           n_fold = 4)
+                           n_fold = 4,
+                           SL_lib = c("SL.mean", "SL.glm", "SL.step")
+                          )
 )
 
 # get origami folds object generated by adaptest to feed into old implementaiton
-folds <- result_new$folds
-folds_vec <- origami::folds2foldvec(folds)
+folds_vec <- origami::folds2foldvec(result_new$folds)
 
 time_old <- system.time(
     result_old <- adaptest:::adaptest_old(Y = Y,
                                           A = A.sample.vec,
                                           n_top = p.true + 5,
                                           n_fold = 4,
-                                          folds_vec = folds_vec)
+                                          folds_vec = folds_vec,
+                                          SL_lib = c("SL.mean", "SL.glm",
+                                                     "SL.step")
+                                         )
 )
 
 # test individual slots in the data_adapt output objects for equivalence
