### 2017-11-21 0ee16f8
move SL deps to suggests; add origami github install for travis/appveyor
