+++ appveyor.yml
@@ -24,6 +24,7 @@ environment:
 
 build_script:
   - travis-tool.sh install_deps
+  - travis-tool.sh install_github jeremyrcoyle/origami
 
 test_script:
   - travis-tool.sh run_tests
