### 2017-11-22 665e208
fix R CMD issue with test re: https://github.com/r-lib/testthat/issues/144
