+++ tests/testthat.R
@@ -1,4 +1,6 @@
 library(testthat)
 library(adaptest)
 
+Sys.setenv(R_TESTS = '')
 test_check("adaptest")
+
