### 2017-12-06 #161 @nhejazi
@karthik Thanks much for checking in. After the initial phase of this review, we began making extensive changes to the software package, both to address review items and for reasons related to the scope of the large research project of which this package is a part. The package has been renamed [`adaptest`](https://github.com/wilsoncai1992/adaptest), and work on the revised version is nearly complete. Our intention is to complete the outstanding review items, submit the package for review to the [Bioconductor project](https://bioconductor.org/), and to continue with this review at that time (we anticipate by end of January 2018). Would this be acceptable?
