+++ NAMESPACE
@@ -1,12 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 S3method(plot,data_adapt)
-S3method(print,data_adapt)
-S3method(summary,data_adapt)
 export(adaptest)
-export(get_composition)
-export(get_pval)
-export(get_significant_biomarker)
 export(rank_DE)
 export(shinyprint.data_adapt)
 importFrom(R2HTML,HTML)
