+++ R/utils.R
@@ -2,8 +2,7 @@
 #'
 #' Customized informative print method for examining data-adaptive statistics
 #'
-#' @param x data-adaptive statistical object of class \code{data_adapt} as
-#'  returned by \code{adaptest}.
+#' @param x (data_adapt) - object of class \code{data_adapt} as returned by \code{adaptest}
 #' @param ... additional arguments passed to \code{print} as necessary
 #'
 #' @keywords internal
@@ -32,8 +31,7 @@ print.data_adapt <- function(x, ...) {
 #'
 #' Provides HTML-based printing utility for examining data-adaptive statistics
 #'
-#' @param x data-adaptive statistical object of class \code{data_adapt}
-#'  as returned by \code{adaptest}
+#' @param x (data_adapt) - object of class \code{data_adapt} as returned by \code{adaptest}
 #'
 #' @importFrom R2HTML HTML
 #'
@@ -52,8 +50,7 @@ shinyprint.data_adapt <- function(x) {
 #'
 #' Customized plotting method for easily examining data-adaptive statistics
 #'
-#' @param x data-adaptive statistical object of class \code{data_adapt} as
-#'  returned by \code{adaptest}
+#' @param x (data_adapt) - object of class \code{data_adapt} as returned by \code{adaptest}
 #' @param plot_type character vector specifying which of the two types of plots
 #'  to generate: "biomarker" for a plot sorted average CV-rank, or "adapt_param"
 #'  for a plot sorted by p-values with labels corresponding to indices
@@ -99,17 +96,20 @@ plot.data_adapt <- function(x, ..., plot_type = c("biomarker", "adapt_param")) {
   }
 }
 
-#' summary of data_adapt
+#' Summary tables of the `adaptest` object
 #'
-#' @param object \code{data_adapt} object
+#' @param object (data_adapt) object as returned by \code{adaptest}
 #' @param ... not implemented
-#' @param type char of either `adapt_param` or `biomarker`.
-#'  `adapt_param` mode outputs summary table of the data adaptive target
-#'  parameter. `biomarker` mode summarizes chracteristis of the biomarkers from
+#' @param type (character) - `adapt_param` or `biomarker`.
+#'  `adapt_param` mode summarizes the data-adaptive target
+#'  parameter. `biomarker` mode summarizes chracteristics of the biomarkers from
 #'  the original data
 #'
-#' @return data.frame of the summary statistics
-#' @keywords internal
+#' @return (data.frame) of the summary statistics
+#' @return \code{type = 'adapt_param'} wtih columns: 'data-adaptive parameters', 'Differential expression', 'p-values', 'q-values'
+#' @return \code{type = 'biomarker'} wtih columns: 'biomakers', 'mean rank', '% appear in top'
+#'
+#' @export
 #' @method summary data_adapt
 #'
 summary.data_adapt <- function(object, ..., type = 'adapt_param') {
