+++ R/adaptest.R
@@ -1,6 +1,6 @@
 #' S3-Style Constructor for Data Adaptive Parameter Class
 #'
-#' @param Y (numeric vector) - continuous or binary outcome variable
+#' @param Y (numeric vector) - continuous or binary biomarkers (outcome variables)
 #' @param A (numeric vector) - binary treatment indicator: \code{1} = treatment, \code{0} = control
 #' @param W (numeric vector, numeric matrix, or numeric data.frame) - matrix of baseline covariates where each column corrspond to one baseline covariate. each row correspond to one observation
 #' @param n_top (integer vector) - value for the number of candidate covariates to generate
@@ -8,7 +8,7 @@
 #' @param n_fold (integer vector) - number of cross-validation folds.
 #' @param absolute (logical) - whether or not to test for absolute effect size. If \code{FALSE}, test for directional effect. This overrides argument \code{negative}.
 #' @param negative (logical) - whether or not to test for negative effect size. If \code{FALSE} = test for positive effect size. This is effective only when \code{absolute = FALSE}.
-#' @param parameter_wrapper (function) - user-defined function that takes input (Y, A, W, absolute, negative) and outputs a (integer vector) containing ranks of outcome variables. For detail, refer to `rank_DE`
+#' @param parameter_wrapper (function) - user-defined function that takes input (Y, A, W, absolute, negative) and outputs a (integer vector) containing ranks of biomarkers (outcome variables). For detail, refer to `rank_DE`
 #' @param SL_lib (character vector) - library of learning algorithms to be used in fitting the "Q" and "g" step of the standard TMLE procedure.
 #'
 #' @return \code{S3} object of class "data_adapt" for data-adaptive multiple
@@ -99,29 +99,32 @@ get_pval <- function(Psi_output, EIC_est_final, alpha = 0.05) {
 
 #' Data-adaptive Statistics for High-Dimensional Multiple Testing
 #'
-#' Performs targeted minimum loss-based estimation (TMLE )of a marginal additive
-#' treatment effect of a binary point treatment on an outcome. The data-adaptive
-#' algorithm is used to perform variable reduction to avoid the disadvantages
-#' associated with multiple testing.
+#' Computes marginal average treatment effects of a binary point treatment on multi-dimensional outcomes, adjusting for baseline covariates, using Targeted Minimum Loss-Based Estimation. A data-mining
+#' algorithm is used to perform biomarker selection before multiple testing to increase power.
 #'
-#' @param Y continuous or binary outcome variable
-#' @param A binary treatment indicator: \code{1} = treatment, \code{0} = control
-#' @param W matrix containing baseline covariates
-#' @param n_top integer value for the number of candidate covariates to generate
+#' @param Y (numeric vector) - continuous or binary biomarkers (outcome variables)
+#' @param A (numeric vector) - binary treatment indicator: \code{1} = treatment, \code{0} = control
+#' @param W (numeric vector, numeric matrix, or numeric data.frame) - matrix of baseline covariates where each column corrspond to one baseline covariate. each row correspond to one observation
+#' @param n_top (integer vector) - value for the number of candidate covariates to generate
 #'  using the data-adaptive estimation algorithm
-#' @param n_fold integer number of folds to be used for cross-validation
-#' @param parameter_wrapper function
-#' @param SL_lib character
-#' @param absolute boolean: \code{TRUE} = test for absolute effect size. This
-#'  \code{FALSE} = test for directional effect. This overrides argument
-#'  \code{negative}.
-#' @param negative boolean: \code{TRUE} = test for negative effect size,
-#'  \code{FALSE} = test for positive effect size
-#' @param p_cutoff A \code{numeric} indicating the cutoff at and below which
-#'  p-values are to be considered significant.
[…]
