+++ vignettes/differentialExpression.Rmd
@@ -1,0 +1,183 @@
+---
+title: "Data-mining biomarkers and perform testing"
+author: "[Wilson Cai](www.stat.berkeley.edu/~wcai/), Nima Hejazi, & Alan Hubbard"
+date: "`r Sys.Date()`"
+output:
+  BiocStyle::html_document
+vignette: >
+  %\VignetteIndexEntry{Data-mining biomarkers and perform testing}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+## Introduction
+
+The `biotmle` R package can be used to isolate biomarkers in two ways: based
+on the associations of genomic objects with an exposure variable of interest.
+In this vignette, we illustrate how to use `biotmle` to isolate and visualize
+genes associated with an __exposure__, using a data set containing microarray
+expression measures from an Illumina platform. In the analysis described below,
+Targeted Minimum Loss-Based Estimation (TMLE) is used to transform the
+microarray expression values based on the influence curve representation of the
+Average Treatment Effect (ATE). Following this transformation, the moderated
+t-statistic of Smyth is used to test for a binary groupwise difference (based
+on the exposure variable), using the tools provided by the R package `limma`.
+
+---
+
+## Biomarker Identification
+
+First, we load the `biotmle` package and the (included) `illuminaData` data set:
+
+```{r setup_data}
+library(dplyr)
+library(biotmle)
+library(biotmleData)
+data(illuminaData)
+suppressMessages(library(SummarizedExperiment))
+"%ni%" = Negate("%in%")
+```
+
+In order to perform Targeted Minimum Loss-Based Estimation, we need three
+separate data structures: (1) _W_, baseline covariates that could potentially
+confound the association of biomarkers with the exposure of interest; (2) _A_,
+the exposure of interest; and (3) _Y_, the biomarkers of interest. All values in
+_W_ and _A_ ought to be discretized, in order to avoid practical violations of
+the assumption of positivity. With the `illuminaData` data set below, we
+discretize the age variable in the phenotype-level data below (this can be
+accessed via the `colData` of the `SummarizedExperiment` object). To invoke the
+biomarker assessment function (`biomarkertmle`), we also need to specify a
+variable of interest (or the position of said variable in the design matrix). We
+do both in just a few lines below:
+
+```{r clean_data}
+# discretize "age" in the phenotype-level data
+colData(illuminaData) <- colData(illuminaData) %>%
+  data.frame %>%
+  dplyr::mutate(age = as.numeric(age > median(age))) %>%
+  DataFrame
+
+# specify column index of treatment/exposure variable of interest
+varInt_index <- which(names(colData(illuminaData)) %in% "benzene")
+```
+
+The TMLE-based biomarker discovery process can be invoked using the
+`biomarkertmle` function. The procedure is quite resource-intensive because it
+evaluates the association of each individual potential biomarker (of which there
+are over 20,000 in the included data set) with an exposure of interest, while
+accounting for potential confounding based on all other covariates included in
+the design matrix. We demonstrate the necessary syntax for calling
+`biomarkertmle` below:
+
+```{r biomarkerTMLE_eval, eval=FALSE}
+biomarkerTMLEout <- biomarkertmle(se = illuminaData,
+                                  varInt = varInt_index,
+                                  family = "gaussian",
+                                  g_lib = c("SL.glmnet", "SL.randomForest",
+                                            "SL.polymars", "SL.mean"),
+                                  Q_lib = c("SL.glmnet", "SL.randomForest",
+                                            "SL.nnet", "SL.mean")
+                                 )
+```
+
+```{r load_biomarkerTMLE_result, echo=FALSE}
+data(biomarkertmleOut)
[…]
