+++ R/data.R
@@ -1,0 +1,13 @@
+
+#' Simulated differential expression data with one exposure
+#'
+#' A dataset containing 1e4 biomarkers and one exposure
+#'
+#' @format A numeric matrix containing 1e4 biomarkers of 1e2 subjects. And a numeric vector containing binary exposures
+#' \describe{
+#'   This is example data to be used in testing the \code{adaptest}
+#'   procedure. Consult the vignettes for how to use this data.
+#' }
+#' @return A matrix \code{Y} and a numeric vector \code{A}.
+#'
+"simpleArray"
