+++ R/data.R
@@ -3,11 +3,24 @@
 #'
 #' A dataset containing 1e4 biomarkers and one exposure
 #'
-#' @format A numeric matrix containing 1e4 biomarkers of 1e2 subjects. And a numeric vector containing binary exposures
+#' @format A numeric matrix containing 1e4 biomarkers of 1e2 subjects.
 #' \describe{
 #'   This is example data to be used in testing the \code{adaptest}
 #'   procedure. Consult the vignettes for how to use this data.
 #' }
-#' @return A matrix \code{Y} and a numeric vector \code{A}.
+#' @return A matrix \code{Y}
 #'
-"simpleArray"
+"Y"
+
+#' Simulated differential expression data with one exposure
+#'
+#' A dataset containing 1e4 biomarkers and one exposure
+#'
+#' @format A numeric vector containing binary exposures
+#' \describe{
+#'   This is example data to be used in testing the \code{adaptest}
+#'   procedure. Consult the vignettes for how to use this data.
+#' }
+#' @return A numeric vector \code{A}.
+#'
+"A"
