+++ DESCRIPTION
@@ -22,6 +22,7 @@ License: GPL-2
 URL: https://github.com/wilsoncai1992/adaptest
 BugReports: https://github.com/wilsoncai1992/adaptest/issues
 Encoding: UTF-8
+LazyData: true
 Imports:
   tmle,
   origami (>= 0.8.2),
