+++ vignettes/differentialExpression.Rmd
@@ -3,7 +3,7 @@ title: "Data-mining biomarkers and perform testing"
 author: "[Wilson Cai](www.stat.berkeley.edu/~wcai/), Nima Hejazi, & Alan Hubbard"
 date: "`r Sys.Date()`"
 output:
-  BiocStyle::html_document
+  # BiocStyle::html_document
 vignette: >
   %\VignetteIndexEntry{Data-mining biomarkers and perform testing}
   %\VignetteEngine{knitr::rmarkdown}
@@ -21,14 +21,12 @@ In this vignette, we illustrate how to use `adaptest` to perform such analysis, 
 
 ## Data-mining stage
 
-First, we load the `adaptest` package and the (included) `illuminaData` data set:
+First, we load the `adaptest` package and the (included) `simpleArray` data set:
 
 ```{r setup_data}
-library(dplyr)
+# library(dplyr)
 library(adaptest)
-library(adaptestData)
-data(illuminaData)
-suppressMessages(library(SummarizedExperiment))
+data(simpleArray)
 "%ni%" = Negate("%in%")
 ```
 
@@ -36,79 +34,40 @@ In order to perform Targeted Minimum Loss-Based Estimation, we need three
 separate data structures: (1) _W_, baseline covariates that could potentially
 confound the association of biomarkers with the exposure of interest; (2) _A_,
 the point exposure of interest; and (3) _Y_, the biomarkers of interest. All values in _A_ ought to be binarized, in order to avoid practical violations of
-the assumption of positivity.
-
-With the `illuminaData` data set below, we
-discretize the age variable in the phenotype-level data below (this can be
-accessed via the `colData` of the `SummarizedExperiment` object). To invoke the
-biomarker assessment function (`biomarkertmle`), we also need to specify a
-variable of interest (or the position of said variable in the design matrix). We
-do both in just a few lines below:
-
-```{r clean_data}
-# discretize "age" in the phenotype-level data
-colData(illuminaData) <- colData(illuminaData) %>%
-  data.frame %>%
-  dplyr::mutate(age = as.numeric(age > median(age))) %>%
-  DataFrame
-
-# specify column index of treatment/exposure variable of interest
-varInt_index <- which(names(colData(illuminaData)) %in% "benzene")
-```
+the assumption of positivity. To invoke the data-adaptive testing function (`adaptest`), we also need to specify the number of top biomarkers `n_top` to the data-mining algorithm, and the number of folds `n_fold` for cross-validation. The smaller `n_top` is, the more selective data-mining algorithm we have. The larger `n_fold` is, more folds are carried our in cross validaiton.
 
 The TMLE-based biomarker discovery process can be invoked using the
-`biomarkertmle` function. The procedure is quite resource-intensive because it
+`adaptest` function. The procedure is quite resource-intensive because it
 evaluates the association of each individual potential biomarker (of which there
-are over 20,000 in the included data set) with an exposure of interest, while
+are 1e3 in the included data set) with an exposure of interest, while
 accounting for potential confounding based on all other covariates included in
 the design matrix. We demonstrate the necessary syntax for calling
-`biomarkertmle` below:
+`adaptest` below:
 
 ```{r biomarkerTMLE_eval, eval=FALSE}
-biomarkerTMLEout <- biomarkertmle(se = illuminaData,
-                                  varInt = varInt_index,
-                                  family = "gaussian",
-                                  g_lib = c("SL.glmnet", "SL.randomForest",
-                                            "SL.polymars", "SL.mean"),
-                                  Q_lib = c("SL.glmnet", "SL.randomForest",
[…]
