+++ vignettes/differentialExpression.Rmd
@@ -24,7 +24,6 @@ In this vignette, we illustrate how to use `adaptest` to perform such analysis, 
 First, we load the `adaptest` package and the (included) `simpleArray` data set:
 
 ```{r setup_data}
-# library(dplyr)
 library(adaptest)
 data(simpleArray)
 "%ni%" = Negate("%in%")
@@ -44,16 +43,17 @@ accounting for potential confounding based on all other covariates included in
 the design matrix. We demonstrate the necessary syntax for calling
 `adaptest` below:
 
-```{r biomarkerTMLE_eval, eval=FALSE}
+```{r adaptest_eval, eval=TRUE}
 adaptestout <- adaptest(Y = Y,
                        A = A,
                        W = NULL,
                        n_top = 10,
-                       n_fold = 3)
+                       n_fold = 3,
+                       parameter_wrapper = adaptest::rank_DE)
 ```
 
-```{r load_biomarkerTMLE_result, eval=FALSE}
-data(adaptestOut)
+```{r load_adaptest_result, eval=FALSE}
+data(adaptestout)
 ```
 
 The output of `adaptest` is an object of class `adaptest`, containing the following objects:
@@ -71,60 +71,26 @@ After invoking `adaptest`, the resultant `adaptest` object will have the slots d
 
 ---
 
-## Visualization of Results
+## Interpret + Visualize the Results
 
-This package provides several plotting methods that can be used to visualize
-the results of the TMLE-based biomarker discovery process.
+This package provides several interpretation methods that can be used to tabular and visualize the results of the data-adaptive tests.
 
-The `plot` method for a `bioTMLE` object will produce a histogram of the
-adjusted p-values of each biomarker (based on the Benjamini-Hochberg procedure
-for controlling the False Discovery Rate) as generated by `limma::topTable`:
+The `get_composition` method for a `adaptest` object will produce a table of composition of each data-adaptive parameters that is significant after multiple testing stage:
 
-```{r pval_hist_limma_adjp}
-# plot(x = limmaTMLEout, type = "pvals_adj")
+```{r get_comp_small}
+get_composition(object = adaptestout, type = 'small')
 ```
 
-Setting the argument `type = "pvals_raw"` will instead produce a histogram of
-the raw p-values _(these are less informative and should, in general, not be
-used for inferential purposes, as the computation producing these p-values
-ignores the multiple testing nature of the biomarker discovery problem)_:
+Setting the argument `type = "big"` will instead produce a table of composition of each data-adaptive parameters before multiple testing stage, so that there are more columns
 
-```{r pval_hist_limma_rawp}
-# plot(x = limmaTMLEout, type = "pvals_raw")
+```{r get_comp_big}
+get_composition(object = adaptestout, type = 'big')
 ```
 
-Heatmaps are useful graphics for visualizing the relationship between measures
-on genomic objects and covariates of interest. The `heatmap_ic` function
-provides this graphic for `bioTMLE` objects, allowing for the relationship
-between the exposure variable and some number of "top" biomarkers (as
-determined by the call to `modtest_ic`) to be visualized. In general, the
-heatmap for `bioTMLE` objects expresses how the contributions of each biomarker
-to the Average Treatment Effect (ATE) vary across differences in the exposure
-variable (that is, there is a causal interpretation to the findings). The plot
-produced is a `ggplot2` object and can be modified in place if stored properly.
-For our analysis:
[…]
