+++ man/print.data_adapt.Rd
@@ -14,4 +14,3 @@
 \description{
 Customized informative print method for examining data-adaptive statistics
 }
-\keyword{internal}
