### 2018-01-24 7445b8a
fix old implementation and broken test
