### 2018-01-24 ca684a9
enforce tidyverse code style
