+++ R/adaptest.R
@@ -24,13 +24,13 @@ data_adapt <- function(Y,
                        parameter_wrapper,
                        SL_lib) {
   if (!is.data.frame(Y)) {
-    if(!is.matrix(Y)) {
+    if (!is.matrix(Y)) {
       stop("argument Y must be a data.frame or a matrix")
     }
     Y <- as.data.frame(Y)
   }
   if (!is.vector(A)) stop("argument A must be numeric")
-  if (!is.null(W)) if(!is.matrix(W)) stop("argument W must be matrix")
+  if (!is.null(W)) if (!is.matrix(W)) stop("argument W must be matrix")
   if (!is.numeric(n_top)) stop("argument n_top must be numeric")
   if (!is.numeric(n_fold)) stop("argument n_fold must be numeric")
   if (!is.logical(absolute)) stop("argument absolute must be boolean/logical")
@@ -48,16 +48,22 @@ data_adapt <- function(Y,
   prob_in_top <- NULL
   folds <- NULL
 
-  out <- structure(list(Y, A, W, n_top, n_fold, absolute, negative,
-                        parameter_wrapper, SL_lib, top_colname, DE,
-                        p_value, q_value, significant_q,
-                        mean_rank_top, prob_in_top, folds),
-                   class = "data_adapt")
+  out <- structure(
+    list(
+      Y, A, W, n_top, n_fold, absolute, negative,
+      parameter_wrapper, SL_lib, top_colname, DE,
+      p_value, q_value, significant_q,
+      mean_rank_top, prob_in_top, folds
+    ),
+    class = "data_adapt"
+  )
 
-  names(out) <- c("Y", "A", "W", "n_top", "n_fold", "absolute", "negative",
-                  "parameter_wrapper", "SL_lib", "top_colname", "DE", "p_value",
-                  "q_value", "significant_q", "mean_rank_top", "prob_in_top",
-                  "folds")
+  names(out) <- c(
+    "Y", "A", "W", "n_top", "n_fold", "absolute", "negative",
+    "parameter_wrapper", "SL_lib", "top_colname", "DE", "p_value",
+    "q_value", "significant_q", "mean_rank_top", "prob_in_top",
+    "folds"
+  )
 
   # export instance of "data_adapt" for downstream use
   return(out)
@@ -89,8 +95,10 @@ get_pval <- function(Psi_output, EIC_est_final, alpha = 0.05) {
   upper <- Psi_output + abs(stats::qnorm(alpha / 2)) * sd_by_col
   lower <- Psi_output - abs(stats::qnorm(alpha / 2)) * sd_by_col
 
-  pval <- stats::pnorm(abs(Psi_output / sd_by_col), mean = 0, sd = 1,
-                       lower.tail = FALSE) * 2
+  pval <- stats::pnorm(
+    abs(Psi_output / sd_by_col), mean = 0, sd = 1,
+    lower.tail = FALSE
+  ) * 2
 
   return(list(pval, upper, lower, sd_by_col))
 }
@@ -139,21 +147,25 @@ adaptest <- function(Y,
                      n_top,
                      n_fold,
                      parameter_wrapper = adaptest::rank_DE,
-                     SL_lib = c("SL.glm", "SL.step", "SL.glm.interaction",
-                                "SL.gam", "SL.earth"),
+                     SL_lib = c(
+                       "SL.glm", "SL.step", "SL.glm.interaction",
+                       "SL.gam", "SL.earth"
+                     ),
                      absolute = FALSE,
                      negative = FALSE,
                      p_cutoff = 0.05,
                      q_cutoff = 0.05) {
 
   # use constructor function to instantiate "data_adapt" object
-  data_adapt <- data_adapt(Y = Y, A = A, W = W,
-                           n_top = n_top,
[…]
