+++ paper/Makefile
@@ -1,6 +1,0 @@
-TITLE = manuscript
-.PHONY : pdf
-
-pdf:
-	pandoc $(TITLE).md -o $(TITLE).pdf \
-		--bibliography $(TITLE).bib
