### 2018-02-01 8c18652
add paper, contributing
