+++ paper/manuscript.bib
@@ -1,0 +1,57 @@
+@article{hubbard2016statistical,
+  title={Statistical Inference for Data Adaptive Target Parameters},
+  author={Hubbard, Alan E and Kherad-Pajouh, Sara and {van der Laan}, Mark J},
+  journal={The International Journal of Biostatistics},
+  volume={12},
+  number={1},
+  pages={3--19},
+  year={2016},
+  doi={10.1515/ijb-2015-0013}
+}
+
+@incollection{hubbard2016mining,
+  title = {Mining with Inference: Data-Adaptive Target Parameters},
+  author = {Hubbard, Alan E and {van der Laan}, Mark J},
+  editor = {Buhlmann, Peter and Drineas, Petros and Kane, Michael and {van der
+    Laan}, Mark J},
+  booktitle = {Handbook of Big Data},
+  publisher = {CRC Press, Taylor \& Francis Group, LLC: Boca Raton, FL},
+  year = {2016}
+}
+
+@article{benjamini1995controlling,
+  title={Controlling the false discovery rate: a practical and powerful approach to multiple testing},
+  author={Benjamini, Yoav and Hochberg, Yosef},
+  journal={Journal of the royal statistical society. Series B (Methodological)},
+  pages={289--300},
+  year={1995},
+  publisher={JSTOR}
+}
+
+@manual{R,
+  title = {R: A Language and Environment for Statistical Computing},
+  author = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address = {Vienna, Austria},
+  year = {2016},
+  url = {https://www.R-project.org/}
+}
+@article{lazzeroni2010cost,
+  title={The cost of large numbers of hypothesis tests on power, effect size and sample size},
+  author={Lazzeroni, LC and Ray, A},
+  journal={Molecular psychiatry},
+  year={2010},
+  publisher={Nature Publishing Group},
+  doi={10.1038/mp.2010.117}
+}
+@article{dunn1961multiple,
+  title={Multiple comparisons among means},
+  author={Dunn, Olive Jean},
+  journal={Journal of the American Statistical Association},
+  volume={56},
+  number={293},
+  pages={52--64},
+  year={1961},
+  publisher={Taylor \& Francis Group},
+  doi={10.2307/2282330}
+}
