+++ paper/Makefile
@@ -1,0 +1,6 @@
+TITLE = manuscript
+.PHONY : pdf
+
+pdf:
+	pandoc $(TITLE).md -o $(TITLE).pdf \
+		--bibliography $(TITLE).bib
