### 2018-02-01 d0ad442
add bioc wrapper sketches
