### 2018-02-01 c4484eb
initial sketch of bioc wrappers
