### 2018-02-01 4b54b69
add bioc reqs
