+++ Makefile
@@ -2,13 +2,14 @@ md:
 	Rscript -e "rmarkdown::render('README.Rmd', output_file = 'README.md')"
 
 site:
+	Rscript -e "rmarkdown::render('README-norefs.Rmd', output_file = 'README.md')"
 	Rscript -e "pkgdown::build_site()"
-
-fastcheck:
-	Rscript -e "devtools::check(build_args = '--no-build-vignettes')"
 
 check:
 	Rscript -e "devtools::check()"
+
+checkfast:
+	Rscript -e "devtools::check(build_args = '--no-build-vignettes')"
 
 bioc:
 	Rscript -e "BiocCheck::BiocCheck('.')"
@@ -22,9 +23,9 @@ doc:
 build:
 	Rscript -e "devtools::build()"
 
+buildfast:
+	Rscript -e "devtools::build(vignettes = FALSE)"
+
 style:
 	Rscript -e "styler::style_pkg()"
-
-cov:
-	Rscript -e "covr::package_coverage(type = 'all', combine_types = FALSE, line_exclusions = list('R/plots.R'))"
 
