### 2018-02-08 65eac69
update docs; add data.table
