+++ R/adaptest.R
@@ -1,18 +1,31 @@
 #' S3-Style Constructor for Data Adaptive Parameter Class
 #'
-#' @param Y (numeric vector) - continuous or binary biomarkers (outcome variables)
-#' @param A (numeric vector) - binary treatment indicator: \code{1} = treatment, \code{0} = control
-#' @param W (numeric vector, numeric matrix, or numeric data.frame) - matrix of baseline covariates where each column corrspond to one baseline covariate. each row correspond to one observation
-#' @param n_top (integer vector) - value for the number of candidate covariates to generate
-#'  using the data-adaptive estimation algorithm
+#' @param Y (numeric vector) - continuous or binary biomarkers outcome variables
+#' @param A (numeric vector) - binary treatment indicator: \code{1} = treatment,
+#'  \code{0} = control
+#' @param W (numeric vector, numeric matrix, or numeric data.frame) - matrix of
+#'  baseline covariates where each column corrspond to one baseline covariate.
+#'  Each row correspond to one observation
+#' @param n_top (integer vector) - value for the number of candidate covariates
+#'  to generate using the data-adaptive estimation algorithm.
 #' @param n_fold (integer vector) - number of cross-validation folds.
-#' @param absolute (logical) - whether or not to test for absolute effect size. If \code{FALSE}, test for directional effect. This overrides argument \code{negative}.
-#' @param negative (logical) - whether or not to test for negative effect size. If \code{FALSE} = test for positive effect size. This is effective only when \code{absolute = FALSE}.
-#' @param parameter_wrapper (function) - user-defined function that takes input (Y, A, W, absolute, negative) and outputs a (integer vector) containing ranks of biomarkers (outcome variables). For detail, refer to `rank_DE`
-#' @param SL_lib (character vector) - library of learning algorithms to be used in fitting the "Q" and "g" step of the standard TMLE procedure.
+#' @param absolute (logical) - whether or not to test for absolute effect size.
+#'  If \code{FALSE}, test for directional effect. This overrides argument
+#'  \code{negative}.
+#' @param negative (logical) - whether or not to test for negative effect size.
+#'  If \code{FALSE} = test for positive effect size. This is effective only when
+#'  \code{absolute = FALSE}.
+#' @param parameter_wrapper (function) - user-defined function that takes input
+#'  (Y, A, W, absolute, negative) and outputs a (integer vector) containing
+#'  ranks of biomarkers (outcome variables). For detail, please refer to the
+#'  documentation for \code{rank_DE}.
+#' @param SL_lib (character vector) - library of learning algorithms to be used
+#'  in fitting the "Q" and "g" step of the standard TMLE procedure.
 #'
 #' @return \code{S3} object of class "data_adapt" for data-adaptive multiple
 #'  testing.
+#'
+#' @importFrom data.table is.data.table as.data.table
 #
 data_adapt <- function(Y,
                        A,
@@ -23,20 +36,22 @@ data_adapt <- function(Y,
                        negative,
                        parameter_wrapper,
                        SL_lib) {
-  if (!is.data.frame(Y)) {
+  if (!data.table::is.data.table(Y)) {
     if (!is.matrix(Y)) {
-      stop("argument Y must be a data.frame or a matrix")
+      stop("Argument Y must be a data.frame or a matrix.")
     }
-    Y <- as.data.frame(Y)
+    Y <- data.table::as.data.table(Y)
   }
-  if (!is.vector(A)) stop("argument A must be numeric")
-  if (!is.null(W)) if (!is.matrix(W)) stop("argument W must be matrix")
-  if (!is.numeric(n_top)) stop("argument n_top must be numeric")
-  if (!is.numeric(n_fold)) stop("argument n_fold must be numeric")
[…]
