+++ vignettes/differentialExpression.Rmd
@@ -1,21 +1,31 @@
 ---
 title: "Data-mining biomarkers and perform testing"
-author: "[Wilson Cai](www.stat.berkeley.edu/~wcai/), Nima Hejazi, & Alan Hubbard"
+author: "[Wilson Cai](www.stat.berkeley.edu/~wcai/),
+[Nima Hejazi](https://nimahejazi.org), and [Alan
+Hubbard](http://hubbard.berkeley.edu/)"
 date: "`r Sys.Date()`"
 output:
   # BiocStyle::html_document
 vignette: >
-  %\VignetteIndexEntry{Data-mining biomarkers and perform testing}
+  %\VignetteIndexEntry{Data-mining Biomarkers and High-Dimensional Testing}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
 
 ## Introduction
 
-The `adaptest` R package can be used to perform data-mining and high-dimensional statistical tests that is common in differential expression studies. The package utilizes a two stage procedure:
-1. data-mining stage: reduce the dimension of biomarkers based on the associations of biomarkers with an exposure variable.
-2. multiple testing stage: adjust for multiple testing to control false positives.
-In this vignette, we illustrate how to use `adaptest` to perform such analysis, using a data set containing microarray expression measures.
+The `adaptest` R package can be used to perform data-mining and high-dimensional
+statistical tests that is common in differential expression studies. The package
+utilizes a two-stage procedure:
+
+1. The _data-mining_ stage: reduce the dimension of biomarkers based on the
+   associations of biomarkers with an exposure variable.
+
+2. The _multiple testing_ stage: adjust for multiple testing to control false
+   positives.
+
+In this vignette, we illustrate how to use `adaptest` to perform such analysis,
+using a data set containing microarray expression measures.
 
 ---
 
@@ -32,62 +42,87 @@ data(simpleArray)
 In order to perform Targeted Minimum Loss-Based Estimation, we need three
 separate data structures: (1) _W_, baseline covariates that could potentially
 confound the association of biomarkers with the exposure of interest; (2) _A_,
-the point exposure of interest; and (3) _Y_, the biomarkers of interest. All values in _A_ ought to be binarized, in order to avoid practical violations of
-the assumption of positivity. To invoke the data-adaptive testing function (`adaptest`), we also need to specify the number of top biomarkers `n_top` to the data-mining algorithm, and the number of folds `n_fold` for cross-validation. The smaller `n_top` is, the more selective data-mining algorithm we have. The larger `n_fold` is, more folds are carried our in cross validaiton.
+the point exposure of interest; and (3) _Y_, the biomarkers of interest. All
+values in _A_ ought to be binarized, in order to avoid practical violations of
+the assumption of positivity. To invoke the data-adaptive testing function
+(`adaptest`), we also need to specify the number of top biomarkers `n_top` to
+the data-mining algorithm, and the number of folds `n_fold` for
+cross-validation. The smaller `n_top` is, the more selective data-mining
+algorithm we have. The larger `n_fold` is, the more folds are carried out in
+the cross-validation procedure.
 
 The TMLE-based biomarker discovery process can be invoked using the
 `adaptest` function. The procedure is quite resource-intensive because it
 evaluates the association of each individual potential biomarker (of which there
-are 1e3 in the included data set) with an exposure of interest, while
+are `r ncol(Y)` in the included data set) with an exposure of interest, while
 accounting for potential confounding based on all other covariates included in
 the design matrix. We demonstrate the necessary syntax for calling
 `adaptest` below:
 
 ```{r adaptest_eval, eval=TRUE}
-adaptestout <- adaptest(Y = Y,
-                       A = A,
-                       W = NULL,
-                       n_top = 10,
-                       n_fold = 3,
[…]
