+++ NAMESPACE
@@ -12,6 +12,8 @@ exportClasses(adapTMLE)
 importClassesFrom(SummarizedExperiment,SummarizedExperiment)
 importFrom(R2HTML,HTML)
 importFrom(calibrate,textxy)
+importFrom(data.table,as.data.table)
+importFrom(data.table,is.data.table)
 importFrom(graphics,abline)
 importFrom(graphics,plot)
 importFrom(magrittr,"%>%")
