+++ NAMESPACE
@@ -3,19 +3,15 @@
 S3method(plot,data_adapt)
 S3method(print,data_adapt)
 S3method(summary,data_adapt)
-export(.adaptmle)
 export(adaptest)
 export(get_composition)
 export(rank_DE)
 export(shinyprint.data_adapt)
-exportClasses(adapTMLE)
-importClassesFrom(SummarizedExperiment,SummarizedExperiment)
 importFrom(R2HTML,HTML)
 importFrom(calibrate,textxy)
 importFrom(graphics,abline)
 importFrom(graphics,plot)
 importFrom(magrittr,"%>%")
-importFrom(methods,setClass)
 importFrom(origami,cross_validate)
 importFrom(origami,make_folds)
 importFrom(origami,training)
