+++ R/adaptest.R
@@ -15,58 +15,58 @@
 #'  testing.
 #
 data_adapt <- function(Y,
-                       A,
-                       W = NULL,
-                       n_top,
-                       n_fold,
-                       absolute,
-                       negative,
-                       parameter_wrapper,
-                       SL_lib) {
-  if (!is.data.frame(Y)) {
-    if (!is.matrix(Y)) {
-      stop("argument Y must be a data.frame or a matrix")
+                                             A,
+                                             W = NULL,
+                                             n_top,
+                                             n_fold,
+                                             absolute,
+                                             negative,
+                                             parameter_wrapper,
+                                             SL_lib) {
+    if (!is.data.frame(Y)) {
+        if (!is.matrix(Y)) {
+            stop("argument Y must be a data.frame or a matrix")
+        }
+        Y <- as.data.frame(Y)
     }
-    Y <- as.data.frame(Y)
-  }
-  if (!is.vector(A)) stop("argument A must be numeric")
-  if (!is.null(W)) if (!is.matrix(W)) stop("argument W must be matrix")
-  if (!is.numeric(n_top)) stop("argument n_top must be numeric")
-  if (!is.numeric(n_fold)) stop("argument n_fold must be numeric")
-  if (!is.logical(absolute)) stop("argument absolute must be boolean/logical")
-  if (!is.logical(negative)) stop("argument negative must be boolean/logical")
-  if (!is.function(parameter_wrapper)) stop("argument parameter_wrapper must be function")
-  if (!is.character(SL_lib)) stop("argument SL_lib must be character")
+    if (!is.vector(A)) stop("argument A must be numeric")
+    if (!is.null(W)) if (!is.matrix(W)) stop("argument W must be matrix")
+    if (!is.numeric(n_top)) stop("argument n_top must be numeric")
+    if (!is.numeric(n_fold)) stop("argument n_fold must be numeric")
+    if (!is.logical(absolute)) stop("argument absolute must be boolean/logical")
+    if (!is.logical(negative)) stop("argument negative must be boolean/logical")
+    if (!is.function(parameter_wrapper)) stop("argument parameter_wrapper must be function")
+    if (!is.character(SL_lib)) stop("argument SL_lib must be character")
 
-  # placeholders for outputs to be included when returning the data_adapt object
-  top_colname <- NULL
-  DE <- NULL
-  p_value <- NULL
-  q_value <- NULL
-  significant_q <- NULL
-  mean_rank_top <- NULL
-  prob_in_top <- NULL
-  folds <- NULL
+    # placeholders for outputs to be included when returning the data_adapt object
+    top_colname <- NULL
+    DE <- NULL
+    p_value <- NULL
+    q_value <- NULL
+    significant_q <- NULL
+    mean_rank_top <- NULL
+    prob_in_top <- NULL
+    folds <- NULL
 
-  out <- structure(
-    list(
-      Y, A, W, n_top, n_fold, absolute, negative,
-      parameter_wrapper, SL_lib, top_colname, DE,
-      p_value, q_value, significant_q,
-      mean_rank_top, prob_in_top, folds
-    ),
-    class = "data_adapt"
-  )
+    out <- structure(
+        list(
+            Y, A, W, n_top, n_fold, absolute, negative,
+            parameter_wrapper, SL_lib, top_colname, DE,
+            p_value, q_value, significant_q,
+            mean_rank_top, prob_in_top, folds
+        ),
+        class = "data_adapt"
+    )
 
-  names(out) <- c(
-    "Y", "A", "W", "n_top", "n_fold", "absolute", "negative",
-    "parameter_wrapper", "SL_lib", "top_colname", "DE", "p_value",
-    "q_value", "significant_q", "mean_rank_top", "prob_in_top",
-    "folds"
-  )
+    names(out) <- c(
[…]
