+++ R/adaptest.R
@@ -15,14 +15,14 @@
 #'  testing.
 #
 data_adapt <- function(Y,
-                                             A,
-                                             W = NULL,
-                                             n_top,
-                                             n_fold,
-                                             absolute,
-                                             negative,
-                                             parameter_wrapper,
-                                             SL_lib) {
+                       A,
+                       W = NULL,
+                       n_top,
+                       n_fold,
+                       absolute,
+                       negative,
+                       parameter_wrapper,
+                       SL_lib) {
     if (!is.data.frame(Y)) {
         if (!is.matrix(Y)) {
             stop("argument Y must be a data.frame or a matrix")
@@ -35,10 +35,12 @@ data_adapt <- function(Y,
     if (!is.numeric(n_fold)) stop("argument n_fold must be numeric")
     if (!is.logical(absolute)) stop("argument absolute must be boolean/logical")
     if (!is.logical(negative)) stop("argument negative must be boolean/logical")
-    if (!is.function(parameter_wrapper)) stop("argument parameter_wrapper must be function")
+    if (!is.function(parameter_wrapper)) stop("argument parameter_wrapper
+                                              must be function")
     if (!is.character(SL_lib)) stop("argument SL_lib must be character")
 
-    # placeholders for outputs to be included when returning the data_adapt object
+    # placeholders for outputs to be included when returning the
+    # data_adapt object
     top_colname <- NULL
     DE <- NULL
     p_value <- NULL
@@ -107,31 +109,57 @@ get_pval <- function(Psi_output, EIC_est_final, alpha = 0.05) {
 
 #' Data-adaptive Statistics for High-Dimensional Multiple Testing
 #'
-#' Computes marginal average treatment effects of a binary point treatment on multi-dimensional outcomes, adjusting for baseline covariates, using Targeted Minimum Loss-Based Estimation. A data-mining
-#' algorithm is used to perform biomarker selection before multiple testing to increase power.
+#' Computes marginal average treatment effects of a binary point treatment on
+#' multi-dimensional outcomes, adjusting for baseline covariates, using Targeted
+#' Minimum Loss-Based Estimation. A data-mining ' algorithm is used to perform
+#' biomarker selection before multiple testing to increase power.
 #'
-#' @param Y (numeric vector) - continuous or binary biomarkers (outcome variables)
-#' @param A (numeric vector) - binary treatment indicator: \code{1} = treatment, \code{0} = control
-#' @param W (numeric vector, numeric matrix, or numeric data.frame) - matrix of baseline covariates where each column corrspond to one baseline covariate. each row correspond to one observation
-#' @param n_top (integer vector) - value for the number of candidate covariates to generate
-#'  using the data-adaptive estimation algorithm
+#' @param Y (numeric vector) - continuous or binary biomarkers
+#' (outcome variables)
+#' @param A (numeric vector) - binary treatment indicator:
+#' \code{1} = treatment, \code{0} = control
+#' @param W (numeric vector, numeric matrix, or numeric data.frame) -
+#'  matrix of baseline covariates where each column corrspond to one baseline
+#'  covariate. each row correspond to one observation
+#' @param n_top (integer vector) - value for the number of candidate covariates
+#'  to generate using the data-adaptive estimation algorithm
 #' @param n_fold (integer vector) - number of cross-validation folds.
-#' @param parameter_wrapper (function) - user-defined function that takes input (Y, A, W, absolute, negative) and outputs a (integer vector) containing ranks of biomarkers (outcome variables). For detail, refer to `rank_DE`
-#' @param SL_lib (character vector) - library of learning algorithms to be used in fitting the "Q" and "g" step of the standard TMLE procedure.
-#' @param absolute (logical) - whether or not to test for absolute effect size. If \code{FALSE}, test for directional effect. This overrides argument \code{negative}.
[…]

[…]
