+++ R/bioadaptest.R
@@ -1,15 +1,39 @@
-bioadaptest <- function(se,
+#' Data Adaptive Multiple Testing for Computational Biology
+#'
+#' A thin wrapper that implements the main data-adaptive multiple hypothesis
+#' testing strategy for data structures commonly found in computational biology
+#' experiments, using the popular SummarizedExperiment container class.
+#'
+#' @param data_in ...
+#' @param var_int ...
+#' @param cntrl_set ...
+#' @param n_top ...
+#' @param n_fold ...
+#' @param parameter_wrapper ...
+#' @param SL_lib ...
+#' @param absolute ...
+#' @param negative ...
+#' @param p_cutoff ...
+#' @param q_cutoff ...
+#'
+#' @return ...
+#'
+#' @export
+#
+bioadaptest <- function(data_in,
                         var_int,
-                        parallel = TRUE,
-                        bppar_type = NULL,
-                        future_param = NULL,
-                        subj_ids = NULL,
-                        g_lib = c("SL.glm", "SL.randomForest", "SL.nnet",
-                                  "SL.polymars", "SL.mean"),
-                        Q_lib = c("SL.glm", "SL.randomForest", "SL.nnet",
-                                  "SL.mean")
-                        ) {
-    # ============================================================================
+                        cntrl_set = NULL,
+                        n_top = 25,
+                        n_fold = 10,
+                        parameter_wrapper = rank_DE,
+                        SL_lib = c("SL.glm", "SL.step", "SL.glm.interaction",
+                                   "SL.gam", "SL.earth", "SL.mean"),
+                        absolute = FALSE,
+                        negative = FALSE,
+                        p_cutoff = 0.05,
+                        q_cutoff = 0.05
+                       ) {
+  # ============================================================================
   # catch input and return in output object for user convenience
   # ============================================================================
   call <- match.call(expand.dots = TRUE)
@@ -18,44 +42,38 @@ bioadaptest <- function(se,
   # invoke S4 class constructor for "adapTMLE" object
   # ============================================================================
   adaptmle <- .adaptmle(
-       SummarizedExperiment(
-          assays = list(expMeasures = assay(se)),
+       SummarizedExperiment::SummarizedExperiment(
+          assays = list(expMeasures = assay(data_in)),
           rowData = rowData(se),
           colData = colData(se)
        ),
        call = call,
-       tmle_out = as.data.frame(matrix(NA, 10, 10)),
+       adaptest_out = as.data.frame(matrix(NA, 10, 10)),
        top_table = as.data.frame(matrix(NA, 10, 10))
   )
 
   # ============================================================================
-  # set up parallelization based on input
+  # wrangle input data structures such that we can feed them to main function
   # ============================================================================
-  doFuture::registerDoFuture()
-  if (parallel == TRUE) {
-    if (!is.null(future_param)) {
-      set_future_param <- parse(text = paste0("future", "::", future_param))
-      future::plan(eval(set_future_param))
-    } else {
-      future::plan(future::multiprocess)
-    }
-  } else if (parallel == FALSE) {
-    warning(paste("Sequential evaluation is strongly discouraged.",
-                  "\n Proceed with caution."))
-    future::plan(future::sequential)
-  }
-  if (!is.null(bppar_type)) {
-    bp_type <- eval(parse(text = paste0("BiocParallel", "::",
-                                        bppar_type, "()")))
-  } else {
-    bp_type <- BiocParallel::DoparParam()
-  }
-  BiocParallel::bpprogressbar(bp_type) <- TRUE
-  BiocParallel::register(bp_type, default = TRUE)
+  W <- cntrl_set
+  A <- var_int
+  Y <- data.table::as.data.table(t(SummarizedExperiment::assay(data_in)))
 
   # ============================================================================
[…]
