+++ R/rank_bootstrap.R
@@ -1,71 +1,71 @@
-composition_out <- get_composition(adaptest_fit, 'big')
-composition_out <- get_composition(adaptest_fit)
-composition_out
-bootstrap_once <- function(data, parameter_wrapper = adaptest::rank_DE) {
-  Y <- data$Y
-  A <- data$A
+# composition_out <- get_composition(adaptest_fit, 'big')
+# composition_out <- get_composition(adaptest_fit)
+# composition_out
+# bootstrap_once <- function(data, parameter_wrapper = adaptest::rank_DE) {
+#   Y <- data$Y
+#   A <- data$A
 
-  # non-parametric bootstrap
-  SAMPLE_PER_BOOTSTRAP <- length(A)
-  indices <- sample(1:length(A), size = SAMPLE_PER_BOOTSTRAP, replace = TRUE) # user specify sample size
-  Y = Y[indices,]
-  A = A[indices]
+#   # non-parametric bootstrap
+#   SAMPLE_PER_BOOTSTRAP <- length(A)
+#   indices <- sample(1:length(A), size = SAMPLE_PER_BOOTSTRAP, replace = TRUE) # user specify sample size
+#   Y = Y[indices,]
+#   A = A[indices]
 
-  # calc rank
-  rank_out <- parameter_wrapper(Y = Y, A = A, W = rep(1, SAMPLE_PER_BOOTSTRAP), absolute = FALSE, negative = FALSE)
-  return(rank_out)
-}
-# =============================================================================
-N_SIMULATION = 1e3
-# N_SIMULATION = 8
-library(foreach)
-# library(Rmpi)
-# library(doMPI)
-# cl = startMPIcluster()
-# registerDoMPI(cl)
-# clusterSize(cl) # just to check
+#   # calc rank
+#   rank_out <- parameter_wrapper(Y = Y, A = A, W = rep(1, SAMPLE_PER_BOOTSTRAP), absolute = FALSE, negative = FALSE)
+#   return(rank_out)
+# }
+# # =============================================================================
+# N_SIMULATION = 1e3
+# # N_SIMULATION = 8
+# library(foreach)
+# # library(Rmpi)
+# # library(doMPI)
+# # cl = startMPIcluster()
+# # registerDoMPI(cl)
+# # clusterSize(cl) # just to check
 
-library(doSNOW)
-library(tcltk)
-nw <- parallel:::detectCores()  # number of workers
-cl <- makeSOCKcluster(nw)
-registerDoSNOW(cl)
+# library(doSNOW)
+# library(tcltk)
+# nw <- parallel:::detectCores()  # number of workers
+# cl <- makeSOCKcluster(nw)
+# registerDoSNOW(cl)
 
-all_CI <- foreach(it2 = 1:(N_SIMULATION),
-                  .combine = rbind,
-                  .packages = c('R6', 'SuperLearner', 'adaptest'),
-                  .inorder = FALSE,
-                  .errorhandling = 'pass',
-                  .verbose = T) %dopar% {
-                    # .verbose = T) %do% {
-                    bootstrap_once(data = simulated, parameter_wrapper = rank_DE)
-                    # bootstrap_once(data = simulated)
-                  }
-head(all_CI)
+# all_CI <- foreach(it2 = 1:(N_SIMULATION),
+#                   .combine = rbind,
+#                   .packages = c('R6', 'SuperLearner', 'adaptest'),
+#                   .inorder = FALSE,
+#                   .errorhandling = 'pass',
+#                   .verbose = T) %dopar% {
[…]
