+++ NAMESPACE
@@ -5,6 +5,7 @@ S3method(print,data_adapt)
 S3method(summary,data_adapt)
 export(adaptest)
 export(get_composition)
+export(get_significant_biomarker)
 export(rank_DE)
 export(shinyprint.data_adapt)
 importFrom(R2HTML,HTML)
