+++ NAMESPACE
@@ -7,6 +7,7 @@ export(adaptest)
 export(get_composition)
 export(get_significant_biomarker)
 export(rank_DE)
+export(rank_ttest)
 export(shinyprint.data_adapt)
 importFrom(R2HTML,HTML)
 importFrom(calibrate,textxy)
