+++ R/utils.R
@@ -9,25 +9,25 @@
 #' @export
 #'
 print.data_adapt <- function(x, ...) {
-    print("The top covariates are")
-    print(get_composition(x, type = "big")[[1]])
-    print("The ATE estiamtes are")
-    print(x$DE)
-    print("The raw p-values are")
-    print(x$p_value)
-    print("The adjusted p-values are")
-    print(x$q_value)
-    print("The top mean CV-rank are (the smaller the better)")
-    print(x$mean_rank_top)
-    print(paste(
-        "The percentage of appearing in top", length(x$top_colname),
-        "are (the larger the better)"
-    ))
-    print(x$prob_in_top * 100)
-    print("The covariates still significant are")
-    print(x$significant_q)
-    print("Their compositions are")
-    print(get_composition(x, type = "small")[[1]])
+  print("The top covariates are")
+  print(get_composition(x, type = "big")[[1]])
+  print("The ATE estiamtes are")
+  print(x$DE)
+  print("The raw p-values are")
+  print(x$p_value)
+  print("The adjusted p-values are")
+  print(x$q_value)
+  print("The top mean CV-rank are (the smaller the better)")
+  print(x$mean_rank_top)
+  print(paste(
+    "The percentage of appearing in top", length(x$top_colname),
+    "are (the larger the better)"
+  ))
+  print(x$prob_in_top * 100)
+  print("The covariates still significant are")
+  print(x$significant_q)
+  print("Their compositions are")
+  print(get_composition(x, type = "small")[[1]])
 }
 
 #' Method of \code{shinyprint} for objects of class \code{data_adapt}
@@ -42,12 +42,12 @@ print.data_adapt <- function(x, ...) {
 #' @export shinyprint.data_adapt
 #'
 shinyprint.data_adapt <- function(x) {
-    print.data_adapt(x)
-    HTML(paste(
-        "<b> The covariates still significant are </b>",
-        paste(x[[5]], collapse = " "),
-        sep = "<br/>"
-    ))
+  print.data_adapt(x)
+  HTML(paste(
+    "<b> The covariates still significant are </b>",
+    paste(x[[5]], collapse = " "),
+    sep = "<br/>"
+  ))
 }
 
 
@@ -68,45 +68,45 @@ shinyprint.data_adapt <- function(x) {
 #' @export
 #
 plot.data_adapt <- function(x, ..., plot_type = c("biomarker", "adapt_param")) {
-    top_index <- x$top_index
-    DE <- x$DE
-    p_value <- x$p_value
-    q_value <- x$q_value
-    significant_q <- x$significant_q
+  top_index <- x$top_index
+  DE <- x$DE
+  p_value <- x$p_value
+  q_value <- x$q_value
+  significant_q <- x$significant_q
 
-    mean_rank_top <- x$mean_rank_top
-    prob_in_top <- x$prob_in_top
-    n_top.want <- length(top_index)
+  mean_rank_top <- x$mean_rank_top
+  prob_in_top <- x$prob_in_top
+  n_top.want <- length(top_index)
 
-    if ("biomarker" %in% plot_type) {
-        # Plot sorted average CV-rank
-        plot(
-            mean_rank_top, ylab = "Mean CV-rank", pch = 20,
-            main = "Mean CV-rank of selected covariates \n (Smaller the better)"
-        )
+  if ("biomarker" %in% plot_type) {
+    # Plot sorted average CV-rank
+    plot(
+      mean_rank_top, ylab = "Mean CV-rank", pch = 20,
+      main = "Mean CV-rank of selected covariates \n (Smaller the better)"
+    )
 
-        calibrate::textxy(
-            (1:n_top.want) - 0.3, mean_rank_top + 0.5, top_index,
[…]
