+++ R/zzz.R
@@ -1,10 +1,6 @@
 .onAttach <- function(...) {
-    packageStartupMessage(paste(
-        "adaptest: Data-adaptive statistics for",
-        "\n high-dimensional multiple testing"
+    packageStartupMessage(paste0(
+        "adaptest v", utils::packageDescription("adaptest")$Version,
+        ": Data-adaptive statistics for high-dimensional multiple testing"
     ))
-    packageStartupMessage(
-        "Version: ",
-        utils::packageDescription("adaptest")$Version
-    )
 }
