+++ R/generate_parameter_wrapper.R
@@ -1,27 +1,19 @@
 #' Compute ranking of biomarkers by sorting effect sizes
 #'
-#' Computes ranking of biomarkers based effect sizes, which are computed by
-#' Targeted Minimum Loss-Based Estimation. This function is designed to be
-#' called inside \code{adaptest}; it should not be run by itself outside of that
-#' context.
+#' Computes ranking of biomarkers based effect sizes, which are computed by Targeted Minimum Loss-Based Estimation. This function
+#' is designed to be called inside \code{adaptest}; it should not be run by
+#' itself outside of that context.
 #'
-#' @param Y (numeric vector) - continuous or binary biomarkers outcome variables
-#' @param A (numeric vector) - binary treatment indicator: \code{1} = treatment,
-#'  \code{0} = control
-#' @param W (numeric vector, numeric matrix, or numeric data.frame) - matrix of
-#'  baseline covariates where each column corrspond to one baseline covariate.
-#'  Each row correspond to one observation
-#' @param absolute (logical) - whether or not to test for absolute effect size.
-#'  If \code{FALSE}, test for directional effect. This overrides argument
-#'  \code{negative}.
-#' @param negative (logical) - whether or not to test for negative effect size.
-#'  If \code{FALSE} = test for positive effect size. This is effective only when
-#'  \code{absolute = FALSE}.
+#' @param Y (numeric vector) - continuous or binary biomarkers (outcome variables)
+#' @param A (numeric vector) - binary treatment indicator: \code{1} = treatment, \code{0} = control
+#' @param W (numeric vector, numeric matrix, or numeric data.frame) - matrix of baseline covariates where each column corrspond to one baseline covariate. each row correspond to one observation
+#' @param absolute (logical) - whether or not to test for absolute effect size. If \code{FALSE}, test for directional effect. This overrides argument \code{negative}.
+#' @param negative (logical) - whether or not to test for negative effect size. If \code{FALSE} = test for positive effect size. This is effective only when \code{absolute = FALSE}.
 #'
 #' @return an \code{integer vector} containing ranks of biomarkers.
 #'
 #' @importFrom tmle tmle
-#' @importFrom stats lm as.formula coef
+#' @importFrom stats lm
 #'
 #' @export
 #
@@ -29,49 +21,46 @@ rank_DE <- function(Y,
                     A,
                     W,
                     absolute = FALSE,
-                    negative = FALSE,
-                    SL_lib = c("SL.glm", "SL.step", "SL.glm.interaction",
-                               "SL.gam", "SL.mean")
-                   ) {
+                    negative = FALSE) {
+  # browser()
   n_here <- nrow(Y)
   p_all <- ncol(Y)
+
   B1_fitted <- rep(0, p_all)
 
-  # coerce into modeling variables
-  A_fit <- as.numeric(A)
-  W_fit <- as.matrix(W)
+  SL_lib <- c("SL.glm", "SL.step", "SL.glm.interaction", "SL.gam")
 
   for (it in seq_len(p_all)) {
-      A_fit <- A
-      Y_fit <- Y[, it]
-      W_fit <- as.matrix(W)
-      # CASE 1: TMLE for DE effect size
-      if (sum(W - as.matrix(rep(1, n_here))) != 0) {
-          # if there are W
-          tmle_fit <- tmle(
-              Y = Y_fit, A = A_fit, W = W,
[…]
