+++ R/adaptest_old.R
@@ -1,7 +1,7 @@
 utils::globalVariables(c(
-    "EIC_est_here", "data_adaptive_index",
-    "index_grid_here", "lower", "p_value", "psi_est_here",
-    "sd_by_col", "upper"
+  "EIC_est_here", "data_adaptive_index",
+  "index_grid_here", "lower", "p_value", "psi_est_here",
+  "sd_by_col", "upper"
 ))
 
 #' OLD Data-Adaptive Algorithm Implementation (for reference only)
@@ -46,193 +46,194 @@ adaptest_old <- function(Y,
                          folds_vec = NULL,
                          parameter_wrapper = adaptest::rank_DE,
                          SL_lib = c(
-                             "SL.glm", "SL.step", "SL.glm.interaction",
-                             "SL.gam", "SL.earth"
+                           "SL.glm", "SL.step", "SL.glm.interaction",
+                           "SL.gam", "SL.earth"
                          ),
                          absolute = FALSE,
                          negative = FALSE) {
-    # use constructor function to instantiate "data_adapt" object
-    data_adapt <- data_adapt(
-        Y = Y, A = A, W = W,
-        n_top = n_top,
-        n_fold = n_fold,
-        absolute = absolute,
-        negative = negative,
-        parameter_wrapper = parameter_wrapper,
-        SL_lib = SL_lib
+
+  # use constructor function to instantiate "data_adapt" object
+  data_adapt <- data_adapt(
+    Y = Y, A = A, W = W,
+    n_top = n_top,
+    n_fold = n_fold,
+    absolute = absolute,
+    negative = negative,
+    parameter_wrapper = parameter_wrapper,
+    SL_lib = SL_lib
+  )
+  # ============================================================================
+  # preparation
+  # ============================================================================
+  n_sim <- nrow(data_adapt$Y)
+  p_all <- ncol(data_adapt$Y)
+
+  # if there is no W input, use intercept as W
+  if (is.null(data_adapt$W)) {
+    W <- as.matrix(rep(1, n_sim))
+    data_adapt$W <- W
+  }
+
+  # ============================================================================
+  # create parameter generating sample
+  # ============================================================================
+  # determine number of samples per fold
+  sample_each_fold <- ceiling(n_sim / n_fold)
+
+  # Generate training set status if not given via an external input
+  if (!is.null(folds_vec)) {
+    index_for_folds <- folds_vec
+  } else {
+    # random index for cross-validation if origami not used
+    index_for_folds <- sample(head(
+      rep(
+        seq_len(n_fold),
+        each = sample_each_fold
+      ),
+      n = n_sim
+    ))
+  }
+
+  # number of observations in each fold
+  table_n_per_fold <- table(index_for_folds)
+
+  rank_in_folds <- matrix(0, nrow = n_fold, ncol = p_all)
+  adapt_param_composition <- matrix(0, nrow = n_fold, ncol = n_top)
+  psi_est_composition <- list()
+  EIC_est_composition <- list()
+  # ============================================================================
+  compute.a.fold <- function(data_adapt, it0) {
+    print(paste("Fold:", it0))
+    chunk_as_est <- it0
+
+    # create parameter generating data
+    Y_param <- data_adapt$Y[index_for_folds != chunk_as_est, ]
+    A_param <- data_adapt$A[index_for_folds != chunk_as_est]
+    W_param <- data_adapt$W[index_for_folds != chunk_as_est, , drop = FALSE]
+
+    # create estimation data
+    Y_est <- data_adapt$Y[index_for_folds == chunk_as_est, ]
+    A_est <- data_adapt$A[index_for_folds == chunk_as_est]
+    W_est <- data_adapt$W[index_for_folds == chunk_as_est, , drop = FALSE]
+
[…]
