+++ man/print.data_adapt.Rd
@@ -7,8 +7,7 @@
 \method{print}{data_adapt}(x, ...)
 }
 \arguments{
-\item{x}{(data_adapt) - object of class \code{data_adapt} as returned by
-\code{adaptest}}
+\item{x}{(data_adapt) - object of class \code{data_adapt} as returned by \code{adaptest}}
 
 \item{...}{additional arguments passed to \code{print} as necessary}
 }
