+++ R/generate_parameter_wrapper.R
@@ -1,22 +1,36 @@
 #' Compute ranking of biomarkers by sorting effect sizes
 #'
-#' Computes ranking of biomarkers based effect sizes, which are computed by Targeted Minimum Loss-Based Estimation. This function
-#' is designed to be called inside \code{adaptest}; it should not be run by
-#' itself outside of that context.
+#' Computes ranking of biomarkers based effect sizes, which are computed by
+#' Targeted Minimum Loss-Based Estimation. This function is designed to be
+#' called inside \code{adaptest}; it should not be run by itself outside of that
+#' context.
 #'
-#' @param Y (numeric vector) - continuous or binary biomarkers (outcome variables)
-#' @param A (numeric vector) - binary treatment indicator: \code{1} = treatment, \code{0} = control
-#' @param W (numeric vector, numeric matrix, or numeric data.frame) - matrix of baseline covariates where each column corrspond to one baseline covariate. each row correspond to one observation
-#' @param absolute (logical) - whether or not to test for absolute effect size. If \code{FALSE}, test for directional effect. This overrides argument \code{negative}.
-#' @param negative (logical) - whether or not to test for negative effect size. If \code{FALSE} = test for positive effect size. This is effective only when \code{absolute = FALSE}.
+#' @param Y (numeric vector) - continuous or binary biomarkers outcome variables
+#' @param A (numeric vector) - binary treatment indicator: \code{1} = treatment,
+#'  \code{0} = control
+#' @param W (numeric vector, numeric matrix, or numeric data.frame) - matrix of
+#'  baseline covariates where each column corrspond to one baseline covariate.
+#'  Each row correspond to one observation
+#' @param absolute (logical) - whether or not to test for absolute effect size.
+#'  If \code{FALSE}, test for directional effect. This overrides argument
+#'  \code{negative}.
+#' @param negative (logical) - whether or not to test for negative effect size.
+#'  If \code{FALSE} = test for positive effect size. This is effective only when
+#'  \code{absolute = FALSE}.
 #'
 #' @return an \code{integer vector} containing ranks of biomarkers.
 #'
 #' @importFrom tmle tmle
-#' @importFrom stats lm
+#' @importFrom stats lm as.formula coef
 #'
 #' @export
-#
+#' @example
+#' set.seed(1234)
+#' library(adaptest)
+#' data(simpleArray)
+#' Y <- Y
+#' A <- A
+#' rank_DE(Y = Y, A = A, W = NULL, absolute = FALSE, negative = FALSE)
 rank_DE <- function(Y,
                     A,
                     W,
@@ -69,16 +83,31 @@ rank_DE <- function(Y,
 
 #' Compute ranking of biomarkers by sorting t-test p-values
 #'
-#' @param Y (numeric vector) - continuous or binary biomarkers (outcome variables)
-#' @param A (numeric vector) - binary treatment indicator: \code{1} = treatment, \code{0} = control
-#' @param W (numeric vector, numeric matrix, or numeric data.frame) - matrix of baseline covariates where each column corrspond to one baseline covariate. each row correspond to one observation
+#' @param Y (numeric vector) - continuous or binary biomarkers outcome variables
+#' @param A (numeric vector) - binary treatment indicator: \code{1} = treatment,
+#'  \code{0} = control
+#' @param W (numeric vector, numeric matrix, or numeric data.frame) - matrix of
+#'  baseline covariates where each column corrspond to one baseline covariate.
+#'  Each row correspond to one observation
+#' @param absolute (logical) - whether or not to test for absolute effect size.
+#'  If \code{FALSE}, test for directional effect. This overrides argument
[…]
