+++ R/get_summary.R
@@ -2,14 +2,32 @@
 #'
 #' Customized informative tables for examining data-adaptive statistics.
 #'
-#' @param object (data_adapt) - object of class \code{data_adapt} as returned by \code{adaptest}
-#' @param type (character) - 'small' or 'big'. 'small' mode returns composition of
-#'  data-adaptive parameters after multiple testing stage. 'big' mode returns composition of
-#'  data-adaptive parameters before multiple testing stage.
+#' @param object (data_adapt) - object of class \code{data_adapt} as returned by
+#'  \code{adaptest}
+#' @param type (character) - 'small' or 'big'. 'small' mode returns composition
+#' of data-adaptive parameters after multiple testing stage. 'big' mode returns
+#' composition of data-adaptive parameters before multiple testing stage.
 #'
-#' @return (numeric matrix) containing what fraction of the data-adaptive parameter comes from which biomarker in the original dataset.
+#' @return (numeric matrix) containing what fraction of the data-adaptive
+#' parameter comes from which biomarker in the original dataset.
 #' @export
+#' @example
+#' set.seed(1234)
+#' library(adaptest)
+#' data(simpleArray)
+#' Y <- Y
+#' A <- A
 #'
+#' adaptest_out <- adaptest(Y = Y,
+#'                          A = A,
+#'                          W = NULL,
+#'                          n_top = 5,
+#'                          n_fold = 3,
+#'                          SL_lib = 'SL.glm',
+#'                          parameter_wrapper = adaptest::rank_DE,
+#'                          absolute = FALSE,
+#'                          negative = FALSE)
+#' get_composition(adaptest_out, type = 'small')
 get_composition <- function(object, type = "small") {
   # browser()
   if (type == "small") col.name <- object$top_colname_significant_q
@@ -41,11 +59,28 @@ get_composition <- function(object, type = "small") {
 }
 
 #' Extract statistically significant biomarkers
-#'
 #' @param object \code{data_adapt} object
+#' @param cutoff cut-off value for composition percentage
 #'
 #' @export
-#
+#'
+#' @examples
+#' set.seed(1234)
+#' library(adaptest)
+#' data(simpleArray)
+#' Y <- Y
+#' A <- A
+#'
+#' adaptest_out <- adaptest(Y = Y,
+#'                          A = A,
+#'                          W = NULL,
+#'                          n_top = 5,
+#'                          n_fold = 3,
+#'                          SL_lib = 'SL.glm',
+#'                          parameter_wrapper = adaptest::rank_DE,
+#'                          absolute = FALSE,
+#'                          negative = FALSE)
+#' get_significant_biomarker(adaptest_out)
 get_significant_biomarker <- function(object, cutoff = .5) {
   if(is.null(get_composition(object, type = "small"))) return(integer()) # catch when nothing is significant
   component_table <- colSums(get_composition(object, type = "small")[[1]])
