+++ vignettes/differentialExpression.Rmd
@@ -66,7 +66,6 @@ adaptest_out <- adaptest(Y = Y,
                          n_top = 5,
                          n_fold = 3,
                          SL_lib = 'SL.glm',
-                         # parameter_wrapper = adaptest::rank_ttest,
                          parameter_wrapper = adaptest::rank_DE,
                          absolute = FALSE,
                          negative = FALSE)
