+++ man/shinyprint.data_adapt.Rd
@@ -10,6 +10,9 @@ shinyprint.data_adapt(x)
 \item{x}{(data_adapt) - object of class \code{data_adapt} as returned by
 \code{adaptest}}
 }
+\value{
+strings of HTML; containing information of the fitted model
+}
 \description{
 Provides HTML-based printing utility for examining data-adaptive statistics
 }
