+++ man/rank_DE.Rd
@@ -33,3 +33,10 @@ Targeted Minimum Loss-Based Estimation. This function is designed to be
 called inside \code{adaptest}; it should not be run by itself outside of that
 context.
 }
+\examples{
+set.seed(1234)
+data(simpleArray)
+Y <- Y
+A <- A
+rank_DE(Y = Y, A = A, W = rep(1, length(A)), absolute = FALSE, negative = FALSE)
+}
