### 2018-02-22 20d31f8
sumexp additions
