### 2018-02-22 fcf0d09
minor fixes for bioc wrappers
