+++ R/adaptest.R
@@ -254,7 +254,6 @@ adaptest <- function(Y,
   # origami folds
   folds <- origami::make_folds(n = n_sim, V = n_fold)
   df_all <- data.frame(Y = Y, A = A, W = W)
-  #browser()
   cv_results <- origami::cross_validate(
     cv_fun = cv_param_est, folds = folds,
     data = df_all,
@@ -270,7 +269,6 @@ adaptest <- function(Y,
   # ============================================================================
   # CV
   # ============================================================================
-  browser()
   rank_in_folds <- matrix(
     data = cv_results$data_adaptive_index, nrow = n_fold,
     ncol = p_all, byrow = TRUE
