### 2018-02-23 fb64198
fix broken plotting index
