+++ vignettes/differentialExpression.Rmd
@@ -63,9 +63,9 @@ matrix. We demonstrate the necessary syntax for calling `adaptest` below:
 adaptest_out <- adaptest(Y = Y,
                          A = A,
                          W = NULL,
-                         n_top = 5,
+                         n_top = 25,
                          n_fold = 3,
-                         SL_lib = 'SL.glm',
+                         SL_lib = c("SL.glm", "SL.mean"),
                          parameter_wrapper = adaptest::rank_DE,
                          absolute = FALSE,
                          negative = FALSE)
