+++ vignettes/differentialExpression.Rmd
@@ -63,8 +63,8 @@ matrix. We demonstrate the necessary syntax for calling `adaptest` below:
 adaptest_out <- adaptest(Y = Y,
                          A = A,
                          W = NULL,
-                         n_top = 25,
-                         n_fold = 3,
+                         n_top = 35,
+                         n_fold = 5,
                          SL_lib = c("SL.glm", "SL.mean"),
                          parameter_wrapper = adaptest::rank_DE,
                          absolute = FALSE,
@@ -139,7 +139,7 @@ object and the popular `airway` data set as an example.
 
 To start, let's load the required packages:
 
-```{r sum_exp, message=FALSE}
+```{r sum_exp, message=FALSE, eval=FALSE}
 library(SummarizedExperiment)
 library(airway)
 data(airway)
@@ -148,7 +148,7 @@ data(airway)
 For simplicity, we'll restrict ourselves to just a random subset of the genes or
 transcripts available from the _airway_ data set.
 
-```{r}
+```{r augment_se, eval=FALSE}
 genes_sub <- order(sample(seq_len(1000)))
 air_reduced <- airway[genes_sub, ]
 ```
@@ -158,7 +158,7 @@ data-intensive procedure, requiring a fairly large sample size. To work with the
 relatively small `airway` data set, we'll simply augment the data by doubling it
 in size in a naive manner:
 
-```{r augment_airway}
+```{r augment_airway, eval=FALSE}
 simple_air <- cbind(air_reduced, air_reduced)
 ```
 
@@ -166,7 +166,7 @@ Now, we can perform the same data-adaptive analysis we discussed above on this
 simplified version of the _airway_ data set, simply by calling the `bioadaptest`
 wrapper function:
 
-```{r}
+```{r bioadaptest, eval=FALSE}
 airway_out <- bioadaptest(data_in = simple_air,
                           var_int = as.matrix(colData(simple_air))[, 3],
                           cntrl_set = NULL,
