### 2018-02-27 4afb708
fix Bioc bindings for SumExp
