+++ NAMESPACE
@@ -19,8 +19,6 @@ importFrom(SummarizedExperiment,assay)
 importFrom(SummarizedExperiment,colData)
 importFrom(SummarizedExperiment,rowData)
 importFrom(calibrate,textxy)
-importFrom(data.table,as.data.table)
-importFrom(data.table,is.data.table)
 importFrom(graphics,abline)
 importFrom(graphics,plot)
 importFrom(magrittr,"%>%")

[…]
