### 2018-02-27 753bfb7
fix CI, deps, and add readme details
