+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: adaptest
 Title: Data-adaptive test statistics for multiple testing in high dimensions
-Version: 0.8.7
+Version: 0.8.8
 Authors@R: c(
     person("Weixin", "Cai", email = "wcai@berkeley.edu",
            role = c("aut", "cre", "cph"),
@@ -44,8 +44,6 @@ Suggests:
     gam,
     nnls,
     airway
-Remotes:
-    github::jeremyrcoyle/origami
 VignetteBuilder: knitr
 RoxygenNote: 6.0.1.9000
 biocViews:
