### 2018-03-06 935e3aa
fix ntop == 1
