+++ R/adaptest.R
@@ -298,12 +298,16 @@ adaptest <- function(Y,
   sd_by_col <- inference_out[[4]]
 
   adaptY_composition <- adapt_param_composition[, seq_len(n_top)]
-  if(class(adaptY_composition) == 'integer') adaptY_composition <- matrix(adaptY_composition, nrow = 1)# catch when n_top == 1; user only want top 1 gene
-  adaptY_composition <- apply(
-    adaptY_composition, 2,
-    function(x) table(x) / sum(table(x))
-  )
-
+  if(class(adaptY_composition) == 'integer') {
+    # catch when n_top == 1; user only want top 1 gene
+    adaptY_composition <- matrix(adaptY_composition, ncol = 1)
+    adaptY_composition <- list(table(adaptY_composition) / sum(table(adaptY_composition)))
+  }else{
+    adaptY_composition <- apply(
+      adaptY_composition, 2,
+      function(x) table(x) / sum(table(x))
+    )
+  }
   # ============================================================================
   # perform FDR correction
   # ============================================================================
