### 2018-03-15 bd4b2f5
fix some irregular `apply` behavior
