+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: adaptest
 Title: Data-adaptive test statistics for multiple testing in high dimensions
-Version: 0.99.0
+Version: 0.99.1
 Authors@R: c(
     person("Weixin", "Cai", email = "wcai@berkeley.edu",
            role = c("aut", "cre", "cph"),
@@ -17,7 +17,7 @@ Description: Data-adaptive test statistics represent a general methodology for
     using the approach of data-adaptive statistical target parameters and
     inference.
 Depends:
-    R (>= 3.4.0)
+    R (>= 3.5.0)
 License: GPL-2
 URL: https://github.com/wilsoncai1992/adaptest
 BugReports: https://github.com/wilsoncai1992/adaptest/issues
