+++ R/utils.R
@@ -34,47 +34,6 @@ print.data_adapt <- function(x, ...) {
 
 ################################################################################
 
-#' Method of \code{shinyprint} for objects of class \code{data_adapt}
-#'
-#' Provides HTML-based printing utility for examining data-adaptive statistics
-#'
-#' @param x (data_adapt) - object of class \code{data_adapt} as returned by
-#' \code{adaptest}
-#'
-#' @return strings of HTML; containing information of the fitted model
-#'
-#' @examples
-#' set.seed(1234)
-#' data(simpleArray)
-#' Y <- Y
-#' A <- A
-#'
-#' adaptest_out <- adaptest(Y = Y,
-#'                          A = A,
-#'                          W = NULL,
-#'                          n_top = 5,
-#'                          n_fold = 3,
-#'                          SL_lib = 'SL.glm',
-#'                          parameter_wrapper = adaptest::rank_DE,
-#'                          absolute = FALSE,
-#'                          negative = FALSE)
-#' # shinyprint.data_adapt(adaptest_out)
-#'
-#' @importFrom R2HTML HTML
-#'
-#' @export shinyprint.data_adapt
-#'
-shinyprint.data_adapt <- function(x) {
-  print.data_adapt(x)
-  HTML(paste(
-    "<b> The covariates still significant are </b>",
-    paste(x[[5]], collapse = " "),
-    sep = "<br/>"
-  ))
-}
-
-################################################################################
-
 #' Plot method for data_adapt objects
 #'
 #' Customized plotting method for easily examining data-adaptive statistics
