+++ R/get_summary.R
@@ -14,11 +14,11 @@
 #' @examples
 #' set.seed(1234)
 #' data(simpleArray)
-#' Y <- Y
-#' A <- A
+#' simulated_array <- simulated_array
+#' simulated_treatment <- simulated_treatment
 #'
-#' adaptest_out <- adaptest(Y = Y,
-#'                          A = A,
+#' adaptest_out <- adaptest(Y = simulated_array,
+#'                          A = simulated_treatment,
 #'                          W = NULL,
 #'                          n_top = 5,
 #'                          n_fold = 3,
@@ -68,11 +68,11 @@ get_composition <- function(object, type = "small") {
 #' @examples
 #' set.seed(1234)
 #' data(simpleArray)
-#' Y <- Y
-#' A <- A
+#' simulated_array <- simulated_array
+#' simulated_treatment <- simulated_treatment
 #'
-#' adaptest_out <- adaptest(Y = Y,
-#'                          A = A,
+#' adaptest_out <- adaptest(Y = simulated_array,
+#'                          A = simulated_treatment,
 #'                          W = NULL,
 #'                          n_top = 5,
 #'                          n_fold = 3,
