+++ man/get_composition.Rd
@@ -24,11 +24,11 @@ Customized informative tables for examining data-adaptive statistics.
 \examples{
 set.seed(1234)
 data(simpleArray)
-Y <- Y
-A <- A
+simulated_array <- simulated_array
+simulated_treatment <- simulated_treatment
 
-adaptest_out <- adaptest(Y = Y,
-                         A = A,
+adaptest_out <- adaptest(Y = simulated_array,
+                         A = simulated_treatment,
                          W = NULL,
                          n_top = 5,
                          n_fold = 3,
