+++ man/shinyprint.data_adapt.Rd
@@ -1,36 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
-\name{shinyprint.data_adapt}
-\alias{shinyprint.data_adapt}
-\title{Method of \code{shinyprint} for objects of class \code{data_adapt}}
-\usage{
-shinyprint.data_adapt(x)
-}
-\arguments{
-\item{x}{(data_adapt) - object of class \code{data_adapt} as returned by
-\code{adaptest}}
-}
-\value{
-strings of HTML; containing information of the fitted model
-}
-\description{
-Provides HTML-based printing utility for examining data-adaptive statistics
-}
-\examples{
-set.seed(1234)
-data(simpleArray)
-Y <- Y
-A <- A
-
-adaptest_out <- adaptest(Y = Y,
-                         A = A,
-                         W = NULL,
-                         n_top = 5,
-                         n_fold = 3,
-                         SL_lib = 'SL.glm',
-                         parameter_wrapper = adaptest::rank_DE,
-                         absolute = FALSE,
-                         negative = FALSE)
-# shinyprint.data_adapt(adaptest_out)
-
-}
