+++ man/simulated_treatment.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/data.R
 \docType{data}
-\name{A}
-\alias{A}
+\name{simulated_treatment}
+\alias{simulated_treatment}
 \title{Simulated differential expression data with one exposure}
 \format{A numeric vector containing binary exposures
 \describe{
@@ -10,10 +10,10 @@
   procedure. Consult the vignettes for how to use this data.
 }}
 \usage{
-A
+simulated_treatment
 }
 \value{
-A numeric vector \code{A}.
+A numeric vector \code{simulated_treatment}.
 }
 \description{
 A dataset containing 1e4 biomarkers and one exposure
