+++ man/rank_ttest.Rd
@@ -33,7 +33,7 @@ Compute ranking of biomarkers by sorting t-test p-values
 \examples{
 set.seed(1234)
 data(simpleArray)
-Y <- Y
-A <- A
-rank_ttest(Y = Y, A = A, W = rep(1, length(A)), absolute = FALSE, negative = FALSE)
+simulated_array <- simulated_array
+simulated_treatment <- simulated_treatment
+rank_ttest(Y = simulated_array, A = simulated_treatment, W = rep(1, length(A)), absolute = FALSE, negative = FALSE)
 }
