+++ man/rank_DE.Rd
@@ -38,5 +38,9 @@ set.seed(1234)
 data(simpleArray)
 simulated_array <- simulated_array
 simulated_treatment <- simulated_treatment
-rank_DE(Y = simulated_array, A = simulated_treatment, W = rep(1, length(A)), absolute = FALSE, negative = FALSE)
+rank_DE(Y = simulated_array,
+        A = simulated_treatment,
+        W = rep(1, length(simulated_treatment)),
+        absolute = FALSE,
+        negative = FALSE)
 }
