+++ R/generate_parameter_wrapper.R
@@ -109,7 +109,11 @@ rank_DE <- function(Y,
 #' data(simpleArray)
 #' simulated_array <- simulated_array
 #' simulated_treatment <- simulated_treatment
-#' rank_ttest(Y = simulated_array, A = simulated_treatment, W = rep(1, length(A)), absolute = FALSE, negative = FALSE)
+#' rank_ttest(Y = simulated_array,
+#'            A = simulated_treatment,
+#'            W = rep(1, length(simulated_treatment)),
+#'            absolute = FALSE,
+#'            negative = FALSE)
 rank_ttest <- function(Y,
                     A,
                     W,
