### 2018-03-21 c9ce649
change `SL_lib` to `learning_library`
