+++ R/adaptest.R
@@ -19,7 +19,7 @@
 #'  (Y, A, W, absolute, negative) and outputs a (integer vector) containing
 #'  ranks of biomarkers (outcome variables). For detail, please refer to the
 #'  documentation for \code{rank_DE}.
-#' @param SL_lib (character vector) - library of learning algorithms to be used
+#' @param learning_library (character vector) - library of learning algorithms to be used
 #'  in fitting the "Q" and "g" step of the standard TMLE procedure.
 #'
 #' @return \code{S3} object of class "data_adapt" for data-adaptive multiple
@@ -33,7 +33,7 @@ data_adapt <- function(Y,
                        absolute,
                        negative,
                        parameter_wrapper,
-                       SL_lib) {
+                       learning_library) {
   if (!is.data.frame(Y)) {
     if (!is.matrix(Y)) {
       stop("argument Y must be a data.frame or a matrix")
@@ -49,7 +49,7 @@ data_adapt <- function(Y,
   if (!is.function(parameter_wrapper)) {
     stop("argument parameter_wrapper must be function")
   }
-  if (!is.character(SL_lib)) stop("argument SL_lib must be character")
+  if (!is.character(learning_library)) stop("argument learning_library must be character")
 
   # placeholders for outputs to be included when returning the data_adapt object
   top_colname <- NULL
@@ -64,7 +64,7 @@ data_adapt <- function(Y,
   out <- structure(
     list(
       Y, A, W, n_top, n_fold, absolute, negative,
-      parameter_wrapper, SL_lib, top_colname, DE,
+      parameter_wrapper, learning_library, top_colname, DE,
       p_value, q_value, significant_q,
       mean_rank_top, prob_in_top, folds
     ),
@@ -73,7 +73,7 @@ data_adapt <- function(Y,
 
   names(out) <- c(
     "Y", "A", "W", "n_top", "n_fold", "absolute", "negative",
-    "parameter_wrapper", "SL_lib", "top_colname", "DE", "p_value",
+    "parameter_wrapper", "learning_library", "top_colname", "DE", "p_value",
     "q_value", "significant_q", "mean_rank_top", "prob_in_top",
     "folds"
   )
@@ -139,7 +139,7 @@ get_pval <- function(Psi_output, EIC_est_final, alpha = 0.05) {
 #'  (Y, A, W, absolute, negative) and outputs a (integer vector) containing
 #'  ranks of biomarkers (outcome variables). For details, please refer to the
 #'  documentation for \code{rank_DE}
-#' @param SL_lib (character vector) - library of learning algorithms to be used
+#' @param learning_library (character vector) - library of learning algorithms to be used
 #'  in fitting the "Q" and "g" step of the standard TMLE procedure.
 #' @param absolute (logical) - whether or not to test for absolute effect size.
 #'  If \code{FALSE}, test for directional effect. This overrides argument
@@ -194,7 +194,7 @@ get_pval <- function(Psi_output, EIC_est_final, alpha = 0.05) {
 #'          W = NULL,
 #'          n_top = 5,
 #'          n_fold = 3,
-#'          SL_lib = 'SL.glm',
+#'          learning_library = 'SL.glm',
 #'          parameter_wrapper = adaptest::rank_DE,
 #'          absolute = FALSE,
 #'          negative = FALSE)
@@ -205,7 +205,7 @@ adaptest <- function(Y,
                      n_top,
                      n_fold,
                      parameter_wrapper = rank_DE,
-                     SL_lib = c(
+                     learning_library = c(
                        "SL.glm", "SL.step", "SL.glm.interaction",
                        "SL.gam", "SL.earth"
                      ),
@@ -226,7 +226,7 @@ adaptest <- function(Y,
     absolute = absolute,
     negative = negative,
[…]
