+++ man/adaptest_old.Rd
@@ -5,8 +5,8 @@
 \title{OLD Data-Adaptive Algorithm Implementation (for reference only)}
 \usage{
 adaptest_old(Y, A, W = NULL, n_top, n_fold, folds_vec = NULL,
-  parameter_wrapper = adaptest::rank_DE, SL_lib = c("SL.glm", "SL.step",
-  "SL.glm.interaction", "SL.gam", "SL.earth"), absolute = FALSE,
+  parameter_wrapper = adaptest::rank_DE, learning_library = c("SL.glm",
+  "SL.step", "SL.glm.interaction", "SL.gam", "SL.earth"), absolute = FALSE,
   negative = FALSE)
 }
 \arguments{
@@ -28,7 +28,7 @@ procedure is used.}
 
 \item{parameter_wrapper}{function}
 
-\item{SL_lib}{character}
+\item{learning_library}{character}
 
 \item{absolute}{boolean: \code{TRUE} = test for absolute effect size. This
 \code{FALSE} = test for directional effect. This overrides argument
