+++ vignettes/differentialExpression.Rmd
@@ -65,7 +65,7 @@ adaptest_out <- adaptest(Y = Y,
                          W = NULL,
                          n_top = 35,
                          n_fold = 5,
-                         SL_lib = c("SL.glm", "SL.mean"),
+                         learning_library = c("SL.glm", "SL.mean"),
                          parameter_wrapper = adaptest::rank_DE,
                          absolute = FALSE,
                          negative = FALSE)
