+++ man/get_composition.Rd
@@ -32,7 +32,7 @@ adaptest_out <- adaptest(Y = simulated_array,
                          W = NULL,
                          n_top = 5,
                          n_fold = 3,
-                         SL_lib = 'SL.glm',
+                         learning_library = 'SL.glm',
                          parameter_wrapper = adaptest::rank_DE,
                          absolute = FALSE,
                          negative = FALSE)

[…]
