### 2018-03-24 9bf90a3
fix sapply / vector allocation; add author details; bump version for PR
