+++ R/adaptest.R
@@ -4,7 +4,7 @@
 #' @param A (numeric vector) - binary treatment indicator: \code{1} = treatment,
 #'  \code{0} = control
 #' @param W (numeric vector, numeric matrix, or numeric data.frame) - matrix of
-#'  baseline covariates where each column corrspond to one baseline covariate.
+#'  baseline covariates where each column correspond to one baseline covariate.
 #'  Each row correspond to one observation
 #' @param n_top (integer vector) - value for the number of candidate covariates
 #'  to generate using the data-adaptive estimation algorithm.
@@ -19,8 +19,8 @@
 #'  (Y, A, W, absolute, negative) and outputs a (integer vector) containing
 #'  ranks of biomarkers (outcome variables). For detail, please refer to the
 #'  documentation for \code{rank_DE}.
-#' @param learning_library (character vector) - library of learning algorithms to be used
-#'  in fitting the "Q" and "g" step of the standard TMLE procedure.
+#' @param learning_library (character vector) - library of learning algorithms
+#'  to be used in fitting the "Q" and "g" step of the standard TMLE procedure.
 #'
 #' @return \code{S3} object of class "data_adapt" for data-adaptive multiple
 #'  testing.
@@ -49,7 +49,9 @@ data_adapt <- function(Y,
   if (!is.function(parameter_wrapper)) {
     stop("argument parameter_wrapper must be function")
   }
-  if (!is.character(learning_library)) stop("argument learning_library must be character")
+  if (!is.character(learning_library)) {
+    stop("argument learning_library must be character")
+  }
 
   # placeholders for outputs to be included when returning the data_adapt object
   top_colname <- NULL
@@ -70,14 +72,12 @@ data_adapt <- function(Y,
     ),
     class = "data_adapt"
   )
-
   names(out) <- c(
     "Y", "A", "W", "n_top", "n_fold", "absolute", "negative",
     "parameter_wrapper", "learning_library", "top_colname", "DE", "p_value",
     "q_value", "significant_q", "mean_rank_top", "prob_in_top",
     "folds"
   )
-
   # export instance of "data_adapt" for downstream use
   return(out)
 }
@@ -112,7 +112,6 @@ get_pval <- function(Psi_output, EIC_est_final, alpha = 0.05) {
     abs(Psi_output / sd_by_col), mean = 0, sd = 1,
     lower.tail = FALSE
   ) * 2
-
   return(list(pval = pval, upper = upper, lower = lower, sd_by_col = sd_by_col))
 }
 
@@ -130,8 +129,8 @@ get_pval <- function(Psi_output, EIC_est_final, alpha = 0.05) {
 #' @param A (numeric vector) - binary treatment indicator:
 #'  \code{1} = treatment, \code{0} = control
 #' @param W (numeric vector, numeric matrix, or numeric data.frame) -
-#'  matrix of baseline covariates where each column corrspond to one baseline
-#'  covariate. each row correspond to one observation
+#'  matrix of baseline covariates where each column correspond to one baseline
+#'  covariate and each row corresponds to one observation.
 #' @param n_top (integer vector) - value for the number of candidate covariates
 #'  to generate using the data-adaptive estimation algorithm
 #' @param n_fold (integer vector) - number of cross-validation folds.
@@ -139,8 +138,8 @@ get_pval <- function(Psi_output, EIC_est_final, alpha = 0.05) {
 #'  (Y, A, W, absolute, negative) and outputs a (integer vector) containing
 #'  ranks of biomarkers (outcome variables). For details, please refer to the
[…]
