+++ man/adaptest.Rd
@@ -17,8 +17,8 @@ adaptest(Y, A, W = NULL, n_top, n_fold, parameter_wrapper = rank_DE,
 \code{1} = treatment, \code{0} = control}
 
 \item{W}{(numeric vector, numeric matrix, or numeric data.frame) -
-matrix of baseline covariates where each column corrspond to one baseline
-covariate. each row correspond to one observation}
+matrix of baseline covariates where each column correspond to one baseline
+covariate and each row corresponds to one observation.}
 
 \item{n_top}{(integer vector) - value for the number of candidate covariates
 to generate using the data-adaptive estimation algorithm}
@@ -30,8 +30,8 @@ to generate using the data-adaptive estimation algorithm}
 ranks of biomarkers (outcome variables). For details, please refer to the
 documentation for \code{rank_DE}}
 
-\item{learning_library}{(character vector) - library of learning algorithms to be used
-in fitting the "Q" and "g" step of the standard TMLE procedure.}
+\item{learning_library}{(character vector) - library of learning algorithms
+to be used in fitting the "Q" and "g" step of the standard TMLE procedure.}
 
 \item{absolute}{(logical) - whether or not to test for absolute effect size.
 If \code{FALSE}, test for directional effect. This overrides argument
