+++ NAMESPACE
@@ -17,6 +17,7 @@ importFrom(SummarizedExperiment,assay)
 importFrom(SummarizedExperiment,colData)
 importFrom(SummarizedExperiment,rowData)
 importFrom(calibrate,textxy)
+importFrom(dplyr,"%>%")
 importFrom(graphics,abline)
 importFrom(graphics,plot)
 importFrom(magrittr,"%>%")
