### 2018-03-25 5ef2398
req dplyr, not magrittr
