+++ vignettes/differentialExpression.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "Data-mining biomarkers and perform testing"
 author: "[Wilson Cai](www.stat.berkeley.edu/~wcai/) and
-[Nima Hejazi](https://nimahejazi.org)"
+  [[Nima Hejazi](https://nimahejazi.org)"
 date: "`r Sys.Date()`"
 output:
   BiocStyle::html_document
@@ -33,7 +33,6 @@ using a data set containing microarray expression measures.
 First, we load the `adaptest` package and the (included) `simpleArray` data set:
 
 ```{r setup_data, message=FALSE}
-# devtools::install_github('wilsoncai1992/adaptest')
 library(adaptest)
 data(simpleArray)
 set.seed(1234)
