### 2018-03-25 1214c3b
fix leftover magrittr call; internalize adaptest_old
