+++ NAMESPACE
@@ -20,7 +20,6 @@ importFrom(calibrate,textxy)
 importFrom(dplyr,"%>%")
 importFrom(graphics,abline)
 importFrom(graphics,plot)
-importFrom(magrittr,"%>%")
 importFrom(methods,new)
 importFrom(methods,setClass)
 importFrom(origami,cross_validate)
