### 2018-03-25 1bc90d5
catch non-SummarizedExperiment case in adaptest
