### 2018-03-26 54e6a45
add accessor; adjust rank_ttest; minor other changes
