+++ R/zzz.R
@@ -1,6 +1,6 @@
 .onAttach <- function(...) {
     packageStartupMessage(paste0(
         "adaptest v", utils::packageDescription("adaptest")$Version,
-        ": Data-adaptive statistics for high-dimensional multiple testing"
+        ": Data-Adaptive Statistics for High-Dimensional Multiple Testing"
     ))
 }
