### 2018-03-26 f1f43d6
address minor BiocCheck items
