+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: adaptest
 Title: Data-Adaptive Statistics for High-Dimensional Multiple Testing
-Version: 0.99.2
+Version: 0.99.3
 Authors@R: c(
     person("Weixin", "Cai", email = "wcai@berkeley.edu",
            role = c("aut", "cre", "cph"),
@@ -21,7 +21,7 @@ Description: Data-adaptive test statistics represent a general methodology for
     methodology to many common data analytic challenges that arise in modern
     computational and genomic biology.
 Depends:
-    R (>= 3.4.0)
+    R (>= 3.5.0)
 License: GPL-2
 URL: https://github.com/wilsoncai1992/adaptest
 BugReports: https://github.com/wilsoncai1992/adaptest/issues
