### 2018-03-28 e4bcffb
rm devel files; fix docs; rm dplyr dep; run stylr
