### 2018-04-02 #161 @wilsoncai1992
Dear @karthik, thanks for your patience with this review. We're very happy to be able to report that we've just recently completed a total re-haul of the R package and have been working with the team at Bioconductor. As of 30 March, the package has been accepted into the upcoming Bioconductor 3.7 release, and we feel now would be a good time to continue with this review. Please let us know what next steps make sense in this situation, as we've made extensive changes to the codebase and many of the items from the earlier review have either been addressed or made irrelevant. Should we start a new review?
