### 2018-04-02 #161 @wilsoncai1992
For reference, this is BIOC review https://github.com/Bioconductor/Contributions/issues/628#issuecomment-373484320
