+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: adaptest
 Title: Data-Adaptive Statistics for High-Dimensional Multiple Testing
-Version: 0.99.4
+Version: 1.0.0
 Authors@R: c(
     person("Weixin", "Cai", email = "wcai@berkeley.edu",
            role = c("aut", "cre", "cph"),
