### 2018-05-02 171a82b
update readme, docs, style, site
