+++ README.Rmd
@@ -42,13 +42,12 @@ parameters, the interested reader is directed to @hubbard2016statistical.
 
 ## Installation
 
-<!--
-For standard use, install from [Bioconductor](https://bioconductor.org):
+For standard use, install from
+[Bioconductor](https://bioconductor.org/packages/adaptest):
 ```{r bioc-installation, eval = FALSE}
 source("https://bioconductor.org/biocLite.R")
 biocLite("adaptest")
 ```
--->
 
 To contribute, install the bleeding-edge _development version_ from GitHub via
 [`devtools`](https://www.rstudio.com/products/rpackages/devtools/):
@@ -57,27 +56,26 @@ To contribute, install the bleeding-edge _development version_ from GitHub via
 devtools::install_github("wilsoncai1992/adaptest")
 ```
 
-<!--
 Current and prior [Bioconductor](https://bioconductor.org) releases are
 available under branches with numbers prefixed by "RELEASE_". For example, to
 install the version of this package available via Bioconductor 3.7, use
 
-```{r bioc-archive-installation, eval = FALSE}
+```{r gh-develop-installation, eval = FALSE}
 devtools::install_github("wilsoncai1992/adaptest", ref = "RELEASE_3_7")
 ```
--->
 
 ---
 
-<!--
 ## Example
 
-This is a basic example which shows you how to solve a common problem:
+For details on how to best use the `methyvim` R package, please consult the most
+recent [package
+vignette](https://bioconductor.org/packages/release/bioc/vignettes/adaptest/inst/doc/differentialExpression.html)
+available through the [Bioconductor
+project](https://bioconductor.org/packages/adaptest).
 
-```{r example}
-## basic example code
-```
--->
+---
+
 ## Issues
 
 If you encounter any bugs or have any specific feature requests, please [file an
@@ -87,10 +85,7 @@ issue](https://github.com/wilsoncai1992/adaptest/issues).
 
 ## Contributions
 
-It is our hope that `adaptest` will grow to be widely adopted as a tool for
-employing data-adaptive multiple hypothesis testing procedures in
-high-dimensional and complex problem settings. To that end, contributions are
-very welcome, though we ask that interested contributors consult our
+Contributions are very welcome. Interested contributors should consult our
 [contribution
 guidelines](https://github.com/wilsoncai1992/adaptest/blob/master/CONTRIBUTING.md)
 prior to submitting a pull request.
@@ -104,7 +99,7 @@ After using the `adaptest` R package, please cite it:
       @article{cai2018adaptest,
         doi = {},
         url = {},
-        year  = {2018},
+        year  = {2018+},
         month = {},
         publisher = {The Open Journal},
         volume = {submitted},
@@ -127,7 +122,7 @@ and [T32 LM012417-02](https://projectreporter.nih.gov/project_info_description.c
 
 ## License
 
-&copy; 2017-2018 [Wilson Cai](https://www.stat.berkeley.edu/~wcai)
+&copy; 2017-2018 [Wilson Cai](https://statistics.berkeley.edu/~wcai)
 
 The software contents of this repository are distributed under the GPL-2
 license. See file `LICENSE` for details.
