+++ .travis.yml
@@ -1,27 +1,24 @@
-language: r
-dist: trusty
-sudo: required
-cache: packages
-warnings_are_errors: true
-
 branches:
   only:
   - master
 
-env:
-  global:
-    - R_BUILD_ARGS="--no-build-vignettes --no-manual"
-    - R_CHECK_ARGS="--no-build-vignettes --no-manual"
+language: r
+sudo: required
+cache: packages
+
+warnings_are_errors: true
+r_build_args: '--no-build-vignettes --no-manual'
+r_check_args: '--no-build-vignettes --no-manual'
 
 r:
-  - release
-  - devel
+  - bioc-release
+  - bioc-devel
 
-r_packages:
-  - devtools
-  - covr
+before_install:
+  - Rscript -e 'update.packages(ask = FALSE)'
 
 r_github_packages:
+  - r-lib/devtools
   - jimhester/covr
   - jeremyrcoyle/origami
 
