+++ .travis.yml
@@ -2,12 +2,16 @@ branches:
   only:
   - master
 
+env:
+  global:
+  - RGL_USE_NULL=TRUE
+
 language: r
 sudo: required
 cache: packages
 
 warnings_are_errors: true
-r_build_args: '--no-build-vignettes --no-manual'
+r_build_args: '--no-manual'
 r_check_args: '--no-build-vignettes --no-manual'
 
 r:
@@ -17,16 +21,20 @@ r:
 before_install:
   - Rscript -e 'update.packages(ask = FALSE)'
 
+r_packages:
+  - devtools
+  - covr
+
 r_github_packages:
   - r-lib/devtools
   - jimhester/covr
-  - jeremyrcoyle/origami
+  - tlverse/origami
 
 bioc_packages:
-  - SummarizedExperiment
-  - BiocParallel
   - BiocStyle
   - BiocCheck
+  - BiocParallel
+  - SummarizedExperiment
   - airway
 
 ## coverage
@@ -37,3 +45,4 @@ notifications:
   email:
     on_success: change
     on_failure: change
+
