### 2018-05-08 3bf8f27
add GenomeInfoDbData and GenomeInfoDb for appveyor
