### 2018-05-19 #161 @nhejazi
@karthik Just checking in — would it make more sense for us to open a new submission issue for this package given the significant changes to it? (Re-naming, Bioconductor acceptance.) If not, please let us know whether you might need anything further from us in order to proceed with the review. Thanks for your help with this.
