### 2018-06-19 #161 @gavinsimpson
FYI; I'm available next week to pick up the review again if you decide to proceed, but after that I am away for a month and unlikely to have time to do a review for 5-6 weeks from now because of conferences/vacation & work commitments when I get back.
