### 2018-06-20 #161 @nhejazi
@karthik -- No problem at all; glad to have your help in the editorial component of this submission now that you're back. To facilitate using this same issue for completion of the review, I've updated the submission comment at the top of the issue to reflect the new name and links to the package/repo (`adaptest` instead of `data.adapt.multi.test`) as well as bumped the appropriate version number (from 0.2.0 to 1.1.0).

@gavinsimpson -- We would be grateful to have your review before you begin your travels. If you have the time to pick up this review again, you'll likely find that many of the previous problems (e.g., lack of unit tests) have been resolved as a consequence of the vetting process used by Bioconductor, though I will gladly dedicate time to making any further changes you point out this/next week.
