### 2018-06-20 5012371
update joss paper + pkgdown fixes
