+++ man/get_results_adaptmle.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/adaptmle.R
+% Please edit documentation in R/adaptmle-class.R
 \name{get_results_adaptmle}
 \alias{get_results_adaptmle}
 \title{Accessor for results of class adaptest}
