+++ paper/Makefile
@@ -1,6 +1,6 @@
-TITLE = manuscript
+TITLE = paper
 .PHONY : pdf
 
 pdf:
-	pandoc $(TITLE).md -o $(TITLE).pdf \
-		--bibliography $(TITLE).bib
+	pandoc $(TITLE).md -o $(TITLE).pdf --bibliography $(TITLE).bib
+
