+++ NAMESPACE
@@ -11,6 +11,7 @@ export(get_significant_biomarker)
 export(rank_DE)
 export(rank_ttest)
 exportClasses(adapTMLE)
+import(S4Vectors)
 importClassesFrom(SummarizedExperiment,SummarizedExperiment)
 importFrom(SummarizedExperiment,SummarizedExperiment)
 importFrom(SummarizedExperiment,assay)

[…]
