### 2018-06-20 5f36432
explicitly import needed classes from S4Vectors
