+++ NAMESPACE
@@ -11,7 +11,8 @@ export(get_significant_biomarker)
 export(rank_DE)
 export(rank_ttest)
 exportClasses(adapTMLE)
-import(S4Vectors)
+importClassesFrom(S4Vectors,Annotated)
+importClassesFrom(S4Vectors,Vector)
 importClassesFrom(SummarizedExperiment,SummarizedExperiment)
 importFrom(SummarizedExperiment,SummarizedExperiment)
 importFrom(SummarizedExperiment,assay)
