### 2018-06-20 wilsoncai1992/adaptest#16 @nhejazi
We believe this has been resolved as of bb90fa5, the latest commit on the version of this package accepted into and released with Bioconductor 3.7, available at https://github.com/wilsoncai1992/adaptest/tree/RELEASE_3_7
