### 2018-08-07 #161 @nhejazi
@gavinsimpson -- Just following up on your potential re-review of this R package. Since it's been ~5-6 weeks since the last comment on this thread, is there a chance you might have time to complete this review of `adaptest` in the near future?
