### 2018-08-07 #161 @wilsoncai1992
Thank you Gavin!

On Tue, Aug 7, 2018 at 2:08 PM Gavin Simpson <notifications@github.com>
wrote:

> Hi, yes, I was a little optimists at being able to do this inbetween my
> trips to Europe. I'm back in the office next week and will complete the
> review then.
>
> On Tue, Aug 7, 2018, 4:03 PM Nima Hejazi, <notifications@github.com>
> wrote:
>
> > @gavinsimpson <https://github.com/gavinsimpson> -- Just following up on
> > your potential re-review of this R package. Since it's been ~5-6 weeks
> > since the last comment on this thread, is there a chance you might have
> > time to complete this review of adaptest in the near future?
> >
> > —
> > You are receiving this because you were mentioned.
> > Reply to this email directly, view it on GitHub
> > <
> https://github.com/openjournals/joss-reviews/issues/161#issuecomment-411201854
> >,
> > or mute the thread
> > <
> https://github.com/notifications/unsubscribe-auth/AAfaiNXQ3ol_2tujEZaB2Vakp5WAX7y1ks5uOgC0gaJpZM4LguZL
> >
> > .
> >
>
