### 2018-08-21 #161 @nhejazi
Hello @gavinsimpson, just another quick follow-up about the review of this R package. We're hoping to complete the review / finalize publication of this software paper in JOSS by the end of the month --- does it appear that your schedule will allow for your review to be completed within this time frame? Thank you in advance for your re-review.
