### 2018-08-28 5c8e9ad
replace BiocInstaller biocLite mentions with BiocManager
