### 2018-10-08 #161 @nhejazi
@arfon - thanks for following up on this. If at all possible, it would be best if we could find a new reviewer and finalize this submission within the next few weeks. Since this R package is part of Bioconductor and their next release (on their biannual schedule) is at the end of this month, having the review of this paper completed co-incident with that release would allow interested/prospective users to make the most of the software and underlying statistical methodology.
