### 2018-10-09 #161 @arfon
@road2stat, @malcolmbarrett - would either of you be willing/able to complete the review of this package for us? @gavinsimpson has made a great start but we could do with some assistance to get across the line here.

Basically we'd be asking you to work through the checklist at the top of this issue and posting your feedback into this review thread.

Thanks in advance for considering!
