### 2018-10-09 #161 @road2stat
@arfon -sorry I'm not an expert in multiple testing or the S4 classes the package builds on. Maybe next time!
