### 2018-10-09 #161 @karthik
I've dashed off a couple of requests. Will update thread in a day or two.
