### 2018-10-11 #161 @kellieotto
@karthik @wilsoncai1992 @nhejazi I made a first pass and put all my comments in issues to the repo. They should be pretty easy fixes.
