### 2018-10-12 #161 @nhejazi
Thanks for the quick turnaround on your review @kellieotto -- much appreciated! We'll start the process of addressing the points you've brought up in each of the issues you've opened.
