### 2018-10-12 c662a79
CI, JOSS, and BioC fixes
