+++ R/adaptmle-class.R
@@ -5,6 +5,7 @@
 #' @importFrom methods setClass new
 #' @importClassesFrom SummarizedExperiment SummarizedExperiment
 #' @importClassesFrom S4Vectors Vector Annotated
+#' @import BiocGenerics
 #'
 #' @export .adaptmle
 #' @exportClass adapTMLE
@@ -85,4 +86,3 @@ get_results_adaptmle <- function(adaptmle_in, data_adapt_out) {
   adaptmle_in@top_index <- data_adapt_out$top_index
   return(adaptmle_in)
 }
-
