+++ README.Rmd
@@ -19,10 +19,13 @@ knitr::opts_chunk$set(
 [![AppVeyor Build  Status](https://ci.appveyor.com/api/projects/status/github/wilsoncai1992/adaptest?branch=master&svg=true)](https://ci.appveyor.com/project/wilsoncai1992/adaptest/)
 [![Coverage Status](https://img.shields.io/codecov/c/github/wilsoncai1992/adaptest/master.svg)](https://codecov.io/github/wilsoncai1992/adaptest?branch=master)
 [![Project Status: WIP – Initial development is in progress, but there has not yet been a stable, usable release suitable for the public.](http://www.repostatus.org/badges/latest/wip.svg)](http://www.repostatus.org/#wip)
+[![BioC status](http://www.bioconductor.org/shields/build/release/bioc/adpatest.svg)](https://bioconductor.org/checkResults/release/bioc-LATEST/adpatest)
+[![Bioc Time](http://bioconductor.org/shields/years-in-bioc/adpatest.svg)](https://bioconductor.org/packages/release/bioc/html/adpatest.html)
+[![Bioc Downloads](http://bioconductor.org/shields/downloads/adpatest.svg)](https://bioconductor.org/packages/release/bioc/html/adpatest.html)
 [![GPL-2 License](http://img.shields.io/:license-gpl2-blue.svg)](http://www.gnu.org/licenses/gpl-2.0.html)
 [![JOSS status](http://joss.theoj.org/papers/7618d7d14ac77f6f502df3f9eac5917d/status.svg)](http://joss.theoj.org/papers/7618d7d14ac77f6f502df3f9eac5917d)
 
-> Data-adaptive statistics for multiple testing in high-dimensional problems
+> Data-adaptive statistics for multiple testing in high-dimensional biology
 
 __Authors:__ [Wilson Cai](https://stat.berkeley.edu/~wcai) and [Nima
 Hejazi](https://nimahejazi.org)
@@ -32,7 +35,7 @@ Hejazi](https://nimahejazi.org)
 ## What's `adaptest`?
 
 The `adaptest` R package is a tool for performing multiple testing on effect
-sizes under high-dimensional settings, using the approach of data-adaptive
+sizes in high-dimensional settings, using the approach of data-adaptive
 statistical target parameters and inference. For technical details on the
 data-adaptive multiple testing procedure, consult @cai2018data. For an
 introduction to statistical inference procedures using data-adaptive target
@@ -52,8 +55,8 @@ if (!("BiocManager" %in% installed.packages())) {
 BiocManager::install("adaptest")
 ```
 
-To contribute, install the bleeding-edge _development version_ from GitHub via
-[`devtools`](https://www.rstudio.com/products/rpackages/devtools/):
+To contribute, install the _development version_ (i.e., branch `master`__) from
+GitHub via [`devtools`](https://www.rstudio.com/products/rpackages/devtools/):
 
 ```{r gh-master-installation, eval = FALSE}
 devtools::install_github("wilsoncai1992/adaptest")
@@ -67,11 +70,15 @@ install the version of this package available via Bioconductor 3.7, use
 devtools::install_github("wilsoncai1992/adaptest", ref = "RELEASE_3_7")
 ```
 
+**_Note_: As the first stable release of this package was through Bioconductor
+v3.7, the minimum version of [R](https://www.r-project.org/) required to install
+`adaptest` is 3.5.0 ("Joy in Playing").**
+
 ---
 
 ## Example
 
-For details on how to best use the `methyvim` R package, please consult the most
+For details on how to best use the `adaptest` R package, please consult the most
 recent [package
 vignette](https://bioconductor.org/packages/release/bioc/vignettes/adaptest/inst/doc/differentialExpression.html)
 available through the [Bioconductor
[…]
