+++ appveyor.yml
@@ -32,14 +32,14 @@ build_script:
   - travis-tool.sh install_github r-lib/devtools
   - travis-tool.sh install_github jimhester/covr
   - travis-tool.sh install_github tlverse/origami
-  - travis-tool.sh install_bioc S4Vectors
   - travis-tool.sh install_bioc BiocCheck
   - travis-tool.sh install_bioc BiocStyle
   - travis-tool.sh install_bioc BiocGenerics
   - travis-tool.sh install_bioc BiocParallel
+  - travis-tool.sh install_bioc S4Vectors
   - travis-tool.sh install_bioc SummarizedExperiment
-  - travis-tool.sh install_bioc GenomeInfoDbData
   - travis-tool.sh install_bioc GenomeInfoDb
+  - travis-tool.sh install_bioc GenomeInfoDbData
   - travis-tool.sh install_bioc airway
   - travis-tool.sh install_bioc_deps
 
@@ -47,7 +47,8 @@ test_script:
   - travis-tool.sh run_tests
 
 on_success:
-  - Rscript -e "covr::codecov(line_exclusions = list('R/plots.R'))"
+  - Rscript -e "covr::codecov(line_exclusions = list('R/utils.R', 'R/get_summary.R', 'R/data.R', 'R/bioadaptest.R'))"
+  - R CMD BiocCheck .
 
 on_failure:
   - 7z a failure.zip *.Rcheck\*
