+++ .travis.yml
@@ -40,7 +40,8 @@ bioc_packages:
 
 ## coverage
 after_success:
-  - Rscript -e "covr::codecov(line_exclusions = list('R/utils.R', 'R/get_summary.R', 'data.R'))"
+  - Rscript -e "covr::codecov(line_exclusions = list('R/utils.R', 'R/get_summary.R', 'R/data.R', 'R/bioadaptest.R'))"
+  - R CMD BiocCheck .
 
 notifications:
   email:

[…]
