### 2018-10-12 7e1da5e
tweak travis/appveyor; rebuild readme
