### 2018-10-12 4e183a9
add sessioninfo pkg for travis; re-add GenomeInfoDb to appveyor
