+++ .travis.yml
@@ -28,6 +28,7 @@ r_packages:
 
 r_github_packages:
   - r-lib/devtools@2d012d1
+  - r-lib/sessioninfo
   - jimhester/covr
   - tlverse/origami
 
