### 2018-10-12 c75674a
more CI tweaks
