+++ NAMESPACE
@@ -11,7 +11,6 @@ export(get_significant_biomarker)
 export(rank_DE)
 export(rank_ttest)
 exportClasses(adapTMLE)
-import(BiocGenerics)
 importClassesFrom(S4Vectors,Annotated)
 importClassesFrom(S4Vectors,Vector)
 importClassesFrom(SummarizedExperiment,SummarizedExperiment)
