### 2018-10-15 9abbc2b
add test for class; fix misspelling; update pkgdown
