### 2018-10-15 92f426f
add vignette details and references
