### 2018-10-15 4de98f9
revise paper based on JOSS review
