+++ paper/paper.md
@@ -21,7 +21,7 @@ authors:
 affiliations:
  - name: Group in Biostatistics, University of California, Berkeley
    index: 1
-date: 20 June 2018
+date: 15 October 2018
 bibliography: manuscript.bib
 ---
 
@@ -41,58 +41,63 @@ Data-adaptive test statistics for multiple testing are motivated by efforts to
 address the limitations of existing multiple testing methods such as the
 popular Benjamini-Hochberg procedure to control the False Discovery Rate (FDR)
 [@benjamini1995controlling] or the Bonferroni method to control the Family-Wise
-Error Rate (FWER) [@dunn1961multiple]. Such methods have been well studied in
-the literature on multiple testing, and it was been well establish that, for a
-fixed targeted effect size and fixed sample size, power decreases as the number
-of tests and corresponding critical values increase [@lazzeroni2010cost].
-Further, @lazzeroni2010cost show that if the power for a single test is 80\%,
-the power is approximately 50\% for 10; 10\% for 1000; and 1\% for 100,000
+Error Rate (FWER) [@dunn1961multiple]. Such methods are well studied in the
+literature on multiple testing, and it is well established that, for a fixed
+targeted effect size and fixed sample size, power decreases as the number of
+tests and corresponding critical values increase [@lazzeroni2010cost]. Further,
+@lazzeroni2010cost show that if the power for a single test is 80\%, the power
+is approximately 50\% for 10; 10\% for 1000; and 1\% for 100,000
 Bonferroni-adjusted tests, a classic method to correct for Type-I error when
 facing multiple testing issues. This simple example demonstrates that data
 analysts and other practitioners must invest, at a prohibitively high rate,
 additional resources to collect samples in order to obtain meaningful results
 under high-dimensional multiple testing constraints.
 
-Utilizing this recently developed data-adaptive statistical framework,
-information loss induced by standard multiple testing procedures can be avoided
-by reducing the dimensionality of problems via data-adaptive variable reduction.
-This recent methodological advance, a data-adaptive multiple testing technique
-[@cai2018data-adaptive], is a natural extension of the data-adaptive target
-parameter framework introduced in @hubbard2016statistical and
-@hubbard2016mining, which present a new class of inference procedures that
-introduce more rigorous statistical inference into problems being increasingly
-addressed by clever yet _ad hoc_ algorithms for data mining.
+Utilizing this recently developed data-adaptive statistical framework, our
+method reduces information loss induced by standard multiple testing procedures
+through data-adaptive dimensionality reduction. This recent methodological
+advance, a data-adaptive multiple testing technique [@cai2018data-adaptive], is
+a natural extension of the data-adaptive target parameter framework introduced
+in @hubbard2016statistical and @hubbard2016mining, which present a new class of
+inference procedures that introduce more rigorous statistical inference into
+problems being increasingly addressed by smart yet _ad hoc_ algorithms for data
+mining.
 
 The approach of data-adaptive test statistics improves on current approaches to
-multiple testing by applying a set of estimation algorithms (specified by the
+multiple testing by applying a set of data-mining algorithms (specified by the
 user) across splits of a particular sample of data, allowing for parameters of
 interest to be discovered from the data. Such methods uncover associations that
 are stable across the full sample and restrict multiple testing to a smaller
 subset of covariates by allowing for variable importance to be measured via the
-data-adaptive procedure. Test statistics formulated in this framework are
-expected to both outperform pre-specified test statistics and provide improved
-power as well as Type I error control, all while simultaneously allowing for
+data-adaptive procedure. Test statistics are formulated on a separately
+held-out subset of data and are expected to both outperform pre-specified test
[…]
