### 2018-10-15 #161 @nhejazi
Great! Many thanks for quickly working through this review with us, @kellieotto!

Many thanks also to @karthik  and @arfon both for searching for a new reviewer and ensuring one was found in time for this review to completed for the upcoming Bioconductor release.
