+++ paper/paper.bib
@@ -18,7 +18,8 @@
   pages={289--300},
   year={1995},
   publisher={JSTOR},
-  url = {http://www.jstor.org/stable/2346101}
+  url = {http://www.jstor.org/stable/2346101},
+  doi={10.2307/2346101}
 }
 
 @manual{R,
