### 2018-10-18 #161 @nhejazi
@karthik --- Done! The DOI is 10.5281/zenodo.1466019, and the Zenodo release link is [here](https://zenodo.org/record/1466019#.W8jEwBRlDD0).

Once again, thanks much for all your help in this process @kellieotto, @karthik, @arfon --- it's much appreciated!
