### 2018-10-18 63c0d26
add zenodo badge, minor readme fixes
