+++ README.Rmd
@@ -18,11 +18,12 @@ knitr::opts_chunk$set(
 [![Travis-CI Build Status](https://travis-ci.org/wilsoncai1992/adaptest.svg?branch=master)](https://travis-ci.org/wilsoncai1992/adaptest?branch=master)
 [![AppVeyor Build  Status](https://ci.appveyor.com/api/projects/status/github/wilsoncai1992/adaptest?branch=master&svg=true)](https://ci.appveyor.com/project/wilsoncai1992/adaptest/)
 [![Coverage Status](https://img.shields.io/codecov/c/github/wilsoncai1992/adaptest/master.svg)](https://codecov.io/github/wilsoncai1992/adaptest?branch=master)
-[![Project Status: WIP – Initial development is in progress, but there has not yet been a stable, usable release suitable for the public.](http://www.repostatus.org/badges/latest/wip.svg)](http://www.repostatus.org/#wip)
+[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![BioC status](http://www.bioconductor.org/shields/build/release/bioc/adaptest.svg)](https://bioconductor.org/checkResults/release/bioc-LATEST/adaptest)
 [![Bioc Time](http://bioconductor.org/shields/years-in-bioc/adaptest.svg)](https://bioconductor.org/packages/release/bioc/html/adaptest.html)
 [![Bioc Downloads](http://bioconductor.org/shields/downloads/adaptest.svg)](https://bioconductor.org/packages/release/bioc/html/adaptest.html)
 [![GPL-2 License](http://img.shields.io/:license-gpl2-blue.svg)](http://www.gnu.org/licenses/gpl-2.0.html)
+[![Zenodo DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1466019.svg)](https://doi.org/10.5281/zenodo.1466019)
 [![JOSS status](http://joss.theoj.org/papers/7618d7d14ac77f6f502df3f9eac5917d/status.svg)](http://joss.theoj.org/papers/7618d7d14ac77f6f502df3f9eac5917d)
 
 > Data-adaptive statistics for multiple testing in high-dimensional biology
@@ -55,8 +56,9 @@ if (!("BiocManager" %in% installed.packages())) {
 BiocManager::install("adaptest")
 ```
 
-To contribute, install the _development version_ (i.e., branch `master`__) from
-GitHub via [`devtools`](https://www.rstudio.com/products/rpackages/devtools/):
+To contribute, install the _development version_ (i.e., branch __`master`__)
+from GitHub via
+[`devtools`](https://www.rstudio.com/products/rpackages/devtools/):
 
 ```{r gh-master-installation, eval = FALSE}
 devtools::install_github("wilsoncai1992/adaptest")
@@ -71,8 +73,8 @@ devtools::install_github("wilsoncai1992/adaptest", ref = "RELEASE_3_7")
 ```
 
 **_Note_: As the first stable release of this package was through Bioconductor
-v3.7, the minimum version of [R](https://www.r-project.org/) required to install
-`adaptest` is 3.5.0 ("Joy in Playing").**
+v3.7, the minimum version of [R](https://www.r-project.org/) required to
+install `adaptest` is 3.5.0 (codename "Joy in Playing").**
 
 ---
 
@@ -109,7 +111,7 @@ After using the `adaptest` R package, please cite the following
       @article{cai2018adaptest,
         doi = {},
         url = {},
-        year  = {2018+},
+        year  = {2018},
         month = {},
         publisher = {The Open Journal},
         volume = {submitted},
