+++ README.md
@@ -9,9 +9,9 @@ Status](https://travis-ci.org/wilsoncai1992/adaptest.svg?branch=master)](https:/
 Status](https://ci.appveyor.com/api/projects/status/github/wilsoncai1992/adaptest?branch=master&svg=true)](https://ci.appveyor.com/project/wilsoncai1992/adaptest/)
 [![Coverage
 Status](https://img.shields.io/codecov/c/github/wilsoncai1992/adaptest/master.svg)](https://codecov.io/github/wilsoncai1992/adaptest?branch=master)
-[![Project Status: WIP – Initial development is in progress, but there
-has not yet been a stable, usable release suitable for the
-public.](http://www.repostatus.org/badges/latest/wip.svg)](http://www.repostatus.org/#wip)
+[![Project Status: Active – The project has reached a stable, usable
+state and is being actively
+developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![BioC
 status](http://www.bioconductor.org/shields/build/release/bioc/adaptest.svg)](https://bioconductor.org/checkResults/release/bioc-LATEST/adaptest)
 [![Bioc
@@ -20,6 +20,8 @@ Time](http://bioconductor.org/shields/years-in-bioc/adaptest.svg)](https://bioco
 Downloads](http://bioconductor.org/shields/downloads/adaptest.svg)](https://bioconductor.org/packages/release/bioc/html/adaptest.html)
 [![GPL-2
 License](http://img.shields.io/:license-gpl2-blue.svg)](http://www.gnu.org/licenses/gpl-2.0.html)
+[![Zenodo
+DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1466019.svg)](https://doi.org/10.5281/zenodo.1466019)
 [![JOSS
 status](http://joss.theoj.org/papers/7618d7d14ac77f6f502df3f9eac5917d/status.svg)](http://joss.theoj.org/papers/7618d7d14ac77f6f502df3f9eac5917d)
 
@@ -57,7 +59,7 @@ BiocManager::install("adaptest")
 ```
 
 To contribute, install the *development version* (i.e., branch
-`master`\_\_) from GitHub via
+**`master`**) from GitHub via
 [`devtools`](https://www.rstudio.com/products/rpackages/devtools/):
 
 ``` r
@@ -76,7 +78,7 @@ devtools::install_github("wilsoncai1992/adaptest", ref = "RELEASE_3_7")
 ***Note*: As the first stable release of this package was through
 Bioconductor v3.7, the minimum version of
 [R](https://www.r-project.org/) required to install `adaptest` is 3.5.0
-(“Joy in Playing”).**
+(codename “Joy in Playing”).**
 
 -----
 
@@ -114,7 +116,7 @@ After using the `adaptest` R package, please cite the following
   @article{cai2018adaptest,
     doi = {},
     url = {},
-    year  = {2018+},
+    year  = {2018},
     month = {},
     publisher = {The Open Journal},
     volume = {submitted},
