### 2018-10-18 #161 @arfon
@nhejazi - could you updated your `paper.md` file to point to `paper.bib` so we can compile the paper please?
